% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopoClean.R
\name{areas_similares}
\alias{areas_similares}
\title{Compare areas}
\usage{
areas_similares(p1, p2, tol = 0.1)
}
\arguments{
\item{p1}{Object of class sf, sfc, sfg or SpatialPolygons}

\item{p2}{Object of class sf, sfc, sfg or SpatialPolygons}

\item{tol}{numeric}
}
\description{
Compare areas
}
\keyword{internal}
