% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_multiprocess.R
\name{close_multiprocess}
\alias{close_multiprocess}
\title{Close parallel process}
\usage{
close_multiprocess(w)
}
\arguments{
\item{w}{numeric. workers or number of cores}
}
\description{
Close parallel process
}
\references{
Revanth Nemani. 2020. https://stackoverflow.com/questions/59652303/kill-futures-from-future-apply-on-linux#comment105470849_59652303
}
