% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connect_grid.R
\name{Connect_grid}
\alias{Connect_grid}
\title{Connectivity indexes in a regular grid}
\usage{
Connect_grid(nodes, region = NULL, grid_pol = NULL, grid_id = NULL,
  grid_type = c("hexagonal", "square"), cellsize = NULL,
  grid_boundary = FALSE, clip = FALSE, tolerance = NULL,
  distance = list(type = "centroid"), probability = 0.5,
  distance_thresholds = NULL, ProtConn = list(attribute =
  "Intersected area", transboundary = NULL), PC_IIC = list(attribute =
  NULL, attribute_weighted_area = FALSE, metric = "PC"), intern = TRUE,
  parallel = FALSE)
}
\arguments{
\item{nodes}{object of class sf, sfc, sfg or SpatialPolygons. nodes shapefile. The shapefile must be in a projected coordinate system.}

\item{region}{object of class sf, sfc, sfg or SpatialPolygons. region shapefile.The shapefile must be in a projected coordinate system.}

\item{grid_pol}{object of class sf, sfc, sfg or SpatialPolygons. Grid hexagones or squares.The shapefile must be in a projected coordinate system.}

\item{grid_id}{character. Column name of the grid ID.}

\item{grid_type}{character. If grid is null you can make a regular grid of "hexagonal" or "square".}

\item{cellsize}{numeric. Grid area (m²).}

\item{grid_boundary}{logical.If TRUE, the Incomplete "hexagons" or "squares" in the boundaries of the region will be discarded}

\item{clip}{logical. If TRUE, the new grid will be clipped to the region area. The operation time will be longer the greater the number of vertices in the polygon of the region, if it is a region with many vertices use the argument "tolerance".}

\item{tolerance}{numeric. If "clip" is equal to TRUE reduces the number of vertices in the region polygon.}

\item{distance}{list.See distancefile(). E.g.: list(type= "centroid", resistance = NULL).}

\item{probability}{numeric. Connection probability to the selected distance threshold, e.g., 0.5 (default) that is 50 percentage of probability connection. Use in case of selecting the "PC" metric or "ProtConn".}

\item{ProtConn}{list. ProtConn parameters: list(attribute = "Intersected area", transboundary = NULL)}

\item{PC_IIC}{list. PC or IIC metric parameters: list(attribute  = NULL, attribute_weighted_area = FALSE, metric = c("IIC", "PC")).}

\item{intern}{logical. Show the progress of the process, default = TRUE.}

\item{parallel}{logical. Parallelize the function using furrr package and multiprocess plan, default = FALSE.}

\item{distance_threshold}{numeric. Distance thresholds to establish connections.}
}
\description{
Use the function to compute the Protected Connected (ProtConn)or EC, PC, IIC indexes on a grid.
}
\examples{
ruta <- system.file("extdata", "WDPA_May2019_MEX-shapefile-polygons.shp", package = "Makurhini")
Protected_areas <- sf::read_sf(ruta)
ruta <- system.file("extdata",  "Region_test.shp", package = "Makurhini")
region <- sf::read_sf(ruta)
hexagons_priority <- Connect_grid(nodes = Protected_areas, region = region, grid_type = "hexagonal", cellsize = 300000000,#100km2
                                  distance = list(type = "centroid"), distance_thresholds = 10000, clip = TRUE,
                                  PC_IIC = list(metric = "PC"), intern = TRUE, parallel = TRUE)
hexagons_priority

hexagons_priority_2 <- Connect_grid(nodes = Protected_areas, region = region, grid_type = "hexagonal", cellsize = 300000000,#100km2
distance = list(type = "centroid"), distance_thresholds = 5000, clip = TRUE, ProtConn = list(attribute ="Intersected area", transboundary = 10000),
intern = TRUE, parallel = TRUE)
hexagons_priority_2
}
\references{
Matt Strimas-Mackey. http://strimas.com/spatial/hexagonal-grids/.\cr
Saura, S., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2017). Protected areas in the world’s ecoregions: How well connected are they? Ecological Indicators, 76, 144–158.\url{https://doi.org/10.1016/j.ecolind.2016.12.047}.\cr
Saura, S. & Torné, J. 2012. Conefor 2.6 user manual (May 2012). Universidad Politécnica de Madrid. Available at \url{www.conefor.org}.\cr
Pascual-Hortal, L. & Saura, S. 2006. Comparison and development of new graph-based landscape connectivity indices: towards the priorization of habitat patches and corridors for conservation. Landscape Ecology 21 (7): 959-967.\cr
Saura, S. & Pascual-Hortal, L. 2007. A new habitat availability index to integrate connectivity in landscape conservation planning: comparison with existing indices and application to a case study. Landscape and Urban Planning 83 (2-3): 91-103.
}
