% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_dPCIIC_links.R
\name{connections_dIIC}
\alias{connections_dIIC}
\title{Estimate dIIC link}
\usage{
connections_dIIC(
  connections = NULL,
  rows = NULL,
  cols = NULL,
  change = NULL,
  dist = NULL,
  distance_threshold = NULL,
  attrib_n = NULL,
  num = NULL,
  graph_n = NULL,
  G1 = 1000,
  min_nodes = 2000,
  parallel = NULL,
  parallel_mode = 0,
  intern = FALSE
)
}
\arguments{
\item{connections}{numeric}

\item{rows}{numeric}

\item{cols}{numeric}

\item{change}{matrix. New distance matrix or connections values}

\item{dist}{matrix. First distance matrix or connections values}

\item{distance_threshold}{numeric}

\item{attrib_n}{numeric. Nodes attribute}

\item{num}{numeric. numPC ID patches = 1}

\item{graph_n}{Graph from igraph package}

\item{G1}{numeric}

\item{min_nodes}{numeric}

\item{parallel}{numeric}

\item{parallel_mode}{numeric}

\item{intern}{logical}
}
\description{
Estimate dIIC link
}
\keyword{internal}
