% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIndex.R
\name{CSIndex}
\alias{CSIndex}
\title{Climate stability index}
\usage{
CSIndex(stable, change, plot = FALSE, reclass = NULL, write = NULL)
}
\arguments{
\item{stable}{RasterLayer. Stable Holdridge Life Zones.}

\item{change}{RasterLayer. Magnitude of change in Holdridge Life Zones.}

\item{plot}{Logical. If TRUE, Climate stability index is plotted.}

\item{reclass}{Vector. Classification system where the maximum value of magnitude of change in Holdridge Life Zones is equal to 0 (Optional). Use it in case the maximum value is greater than 8, e.g., if maximum value = 9, then: reclass=c(-Inf,0,NA, 0,1,9, 1,2,8, 2,3,7, 3,4,6, 4,5,5, 5,6,4, 6,7,3, 7,8,2, 8,Inf,1).}

\item{write}{Character. Write the output raster e.g., "C:/folder/raster.tif"}
}
\value{
Raster with values from 0 to 1, where 1 represents a zone of stable climate and decreases to 0 as the climate change is greater.
}
\description{
Climate stability index
}
