% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dECAfun.R
\name{dECAfun}
\alias{dECAfun}
\title{Function to contrast differences of dA and dECA}
\usage{
dECAfun(dECA, dA)
}
\arguments{
\item{dECA}{dECA}

\item{dA}{dA}
}
\description{
Function to contrast differences of dA and dECA
}
\keyword{internal}
