% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_ProtConn.R
\name{delta_ProtConn}
\alias{delta_ProtConn}
\title{Delta ProtConn}
\usage{
delta_ProtConn(x = NULL, y = NULL, base_param3 = NULL)
}
\arguments{
\item{x}{object of class sf, sfc, sfg or SpatialPolygons. The file must have a projected coordinate system.}

\item{y}{object of class sf, sfc, sfg or SpatialPolygons. The file must have a projected coordinate system.}

\item{base_param3}{list}
}
\description{
Estimate the contribution of each node to the ProtConn value in the region.
}
