% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_ProtConn_raster.R
\name{delta_ProtConn_raster}
\alias{delta_ProtConn_raster}
\title{Delta ProtConn}
\usage{
delta_ProtConn_raster(
  x = NULL,
  y = NULL,
  distance_thresholds = NULL,
  probability = 0.5,
  LA = NULL,
  distance = NULL,
  resist = NULL,
  works = NULL
)
}
\arguments{
\item{x}{object of class sf, sfc, sfg or SpatialPolygons. The file must have a projected coordinate system.}

\item{y}{object of class sf, sfc, sfg or SpatialPolygons. The file must have a projected coordinate system.}

\item{distance_thresholds}{numeric}

\item{probability}{numeric}

\item{distance}{list}

\item{resist}{raster}

\item{works}{numeric}
}
\description{
Estimate the contribution of each node to the ProtConn value in the region.
}
\keyword{internal}
