% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dist_restoration}
\alias{dist_restoration}
\title{Distance matrix between habitat nodes in Spain after an hypothetical restoration}
\format{
A \code{matrix} with 359 files and 359 columns.
}
\usage{
dist_restoration
}
\description{
A \code{Large matrix} with 128881 elements or 359 files and 359 columns. The distance file gives the potential (after hypothetical restoration measures) effective distance of corridors between patches.
}
\references{
Goicolea, T., Cisneros-Araújo, P., Vega, C.A. et al. Landscape connectivity for predicting the spread of ASF in the European wild boar population. Sci Rep 14, 3414 (2024). https://doi.org/10.1038/s41598-024-53869-5
}
\keyword{datasets}
