% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ecoregions}
\alias{Ecoregions}
\title{Ecoregions vector for analysis}
\format{
A \code{SpatialPolygonsDataFrame} with 3 features and 3 fields, which are:
}
\usage{
Ecoregions
}
\description{
A simple feature collection \code{sf}  with 33 ecoregions of Colombia.
}
\references{
Dinerstein, E., Olson, D., Joshi, A., Vynne, C., Burgess, N. D., Wikramanayake, E., and Kindt, R. (2017).
An Ecoregion-Based Approach to Protecting Half the Terrestrial Realm. BioScience, 67(6), 534?545. https://doi.org/10.1093/biosci/bix014
https://ecoregions2017.appspot.com
}
\keyword{datasets}
