% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXTraster.R
\name{EXTraster}
\alias{EXTraster}
\title{Function to have two raster with the same number of rows, columns and extension}
\usage{
EXTraster(R.Emax, R.Emin)
}
\arguments{
\item{R.Emax}{RasterLayer. Raster with the largest extension. The extent of this raster will be transformed.}

\item{R.Emin}{RasterLayer. Raster with the least extension.}
}
\description{
Function to have two raster with the same number of rows, columns and extension
}
\keyword{internal}
