% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid.R
\name{get_grid}
\alias{get_grid}
\title{ProtConn: input files}
\usage{
get_grid(
  region = NULL,
  grid_pol = NULL,
  grid_id = NULL,
  hexagonal = TRUE,
  cellsize = NULL,
  grid_boundary = FALSE,
  clip = FALSE,
  tolerance = NULL
)
}
\arguments{
\item{region}{object of class sf, sfc, sfg or SpatialPolygons}

\item{grid_pol}{object of class sf, sfc, sfg or SpatialPolygons. Grid hexagones or squares. The
shapefile must be in a projected coordinate system.}

\item{grid_id}{character. Column name of the grid ID.}

\item{hexagonal}{logical. If FALSE will be a regular grid of "square".}

\item{cellsize}{numeric. Grid area (square kilometers).}

\item{grid_boundary}{logical.If TRUE, the Incomplete "hexagons" or "squares" in the boundaries of
the region will be discarded}

\item{clip}{logical. If TRUE, the new grid will be clipped to the region area. The operation time
will be longer the greater the number of vertices in the polygon of the region,
if it is a region with many vertices use the argument "tolerance".}

\item{tolerance}{numeric. If "clip" is equal to TRUE reduces the number of vertices in the region polygon.}
}
\description{
ProtConn: input files
}
