% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pc_grid.R
\name{get_pc_grid}
\alias{get_pc_grid}
\title{Estimates PC metric}
\usage{
get_pc_grid(x, y, p, pmedian = TRUE, d, LA = NULL)
}
\arguments{
\item{x}{list.}

\item{y}{distance matrix}

\item{p}{numeric. Probability}

\item{pmedian}{logical. median (TRUE) or mean(FALSE) dispersal distance}

\item{d}{numeric. Dispersal distance}

\item{LA}{numeric. Max. landscape attribute}
}
\description{
Estimates PC metric
}
\keyword{internal}
