% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sdist.R
\name{get_sdist}
\alias{get_sdist}
\title{Estimate all shortest distance between nodes using weights}
\usage{
get_sdist(
  dist_nodes = NULL,
  graph_nodes = NULL,
  attr_nodes = NULL,
  metric = NULL,
  probability = NULL,
  distance_threshold = NULL,
  igraph_Dijkstra = FALSE,
  parallel = NULL,
  loop = TRUE,
  G1 = 1000,
  min_nodes = 2000,
  return_graph = FALSE,
  intern = TRUE
)
}
\arguments{
\item{dist_nodes}{matrix distance or pairwise data.frame distance}

\item{attr_nodes}{numeric. Nodes attribute}

\item{metric}{character. 'PC' or 'IIC'}

\item{probability}{numeric}

\item{distance_threshold}{numeric}

\item{igraph_Dijkstra}{logical. Use or not igraph to implement the Dijkstra Algorithm}

\item{parallel}{numeric}

\item{loop}{logical}

\item{G1}{numeric}

\item{min_nodes}{numeric}

\item{intern}{logical}
}
\description{
Estimate all shortest distance between nodes using weights
}
\keyword{internal}
