% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{habitat_nodes_raster}
\alias{habitat_nodes_raster}
\title{habitat nodes in raster format}
\format{
A \code{raster} with 404 integer values.
\describe{
\item{id}{habitat patches identifier.}
 }
}
\usage{
habitat_nodes_raster
}
\description{
A simple feature collection \code{sf} with 404 habitat patches/nodes with less alteration or degradation that
 is for testing and running of MK_fragmentation and MK_dPCIIC examples.
}
\references{
Correa Ayram, C. A., Mendoza, M. E., Etter, A., & Pérez Salicrup, D. R. (2017). Anthropogenic impact on habitat connectivity: A multidimensional human footprint index evaluated in a highly biodiverse landscape of Mexico. Ecological Indicators, 72, 895-909. https://doi.org/10.1016/j.ecolind.2016.09.007
}
\keyword{datasets}
