% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HFP_Mexico.R
\docType{data}
\name{HFP_Mexico}
\alias{HFP_Mexico}
\title{Human footprint of Mexico}
\format{
A \code{raster} with Human footprint values from 0 to 100:
\describe{
\item{values}{Human footprint values from 0 to 100}
 }
}
\usage{
HFP_Mexico
}
\description{
A RasterLayer with Human footprint values.
}
\references{
Wildlife Conservation Society - WCS, and Center for International Earth Science Information Network - CIESIN - Columbia University.
2005. Last of the Wild Project, Version 2, 2005 (LWP-2): Global Human Footprint Dataset (Geographic). Palisades, NY: NASA Socioeconomic Data and Applications Center (SEDAC). https://doi.org/10.7927/H4M61H5F.
}
\keyword{datasets}
