% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HQuality.R
\name{HQuality}
\alias{HQuality}
\title{Habitat quality weighted by node area and estandarized}
\usage{
HQuality(x, y, id = NULL, area_unit = NULL, by_zone = NULL,
  write = NULL, SAGA = TRUE)
}
\arguments{
\item{x}{RasterLayer. Habitat quality raster. Proyected coordinate system equal to y shapefile.}

\item{y}{Nodes of class sf, sfc, sfg or SpatialPolygons. Proyected coordinate system equal to x.}

\item{id}{Character. Column name with the nodes id.}

\item{area_unit}{numeric. Area unit to weighte the habitat quality index, udunits2 package compatible unit (e.g., "km2", "cm2", "ha"). If NULL, then it will be equal to square meters "m2".}

\item{by_zone}{Character. Colname with the zones to which nodes belongs. A shapefile will be exported by zone (e.g., by ANP, by Ecoregion).}

\item{write}{Character. Provide folder, name and extension of the output shape, e.g., "C:/folder/CSI.shp". If byzone is not NULL you have to provide only the folder output, e.g., "C:/out/".}

\item{SAGA}{Logical. Optimize the process using SAGA GIS and RSAGA package (see, \url{https://github.com/r-spatial/RSAGA}).}
}
\value{
SpatialPolygons with value of the climatic stability index weighted by the area, standardized and not standardized for all the nodes and by area of interest.
}
\description{
Habitat quality weighted by node area and estandarized
}
