% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_grid.R
\name{input_grid}
\alias{input_grid}
\title{ProtConn: input files}
\usage{
input_grid(
  node,
  landscape = NULL,
  unit = "ha",
  bdist = NULL,
  xsimplify = FALSE
)
}
\arguments{
\item{node}{object of class sf, sfc, sfg or SpatialPolygons}

\item{landscape}{object of class sf, sfc, sfg or SpatialPolygons}

\item{unit}{character}

\item{bdist}{numeric}

\item{xsimplify}{logical or numeric}
}
\description{
ProtConn: input files
}
