% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation_patches.R
\docType{data}
\name{list_forest_patches}
\alias{list_forest_patches}
\title{List of vegetation patches}
\format{
A \code{sf} with 4 shapefiles:
}
\usage{
list_forest_patches
}
\description{
A list with 4 shapefiles. Each shapefile correspond to vegetation patches with
less alteration or degradation  in the region of the Biosphere Reserve Mariposa Monarca for
the nex periods: "1993", "2003", "2007", "2011". The list is used for testing and running
of MK_dECA and test_ECA_distance examples.
}
\references{
INEGI. (2001). Conjunto de datos vectoriales de uso del suelo y vegetacion, serie II (Continuo nacional), escala 1:250, 000.
Instituto Nacional de Estadistica y Geografia, Aguascalientes.
INEGI. (2005). Conjunto de datos vectoriales de uso del suelo y vegetacion, serie III (Continuo nacional), escala 1:250, 000.
Instituto Nacional de Estadistica y Geografia, Aguascalientes.
INEGI. (2009). Conjunto de datos vectoriales de uso del suelo y vegetacion, serie IV (Continuo nacional), escala 1:250, 000.
Instituto Nacional de Estadistica y Geografia, Aguascalientes.
INEGI. (2013). Conjunto de datos vectoriales de uso del suelo y vegetacion, serie V (capa union), escala 1:250, 000.
Instituto Nacional de Estadistica y Geografia, Aguascalientes.
}
\keyword{datasets}
