% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\title{Make grid}
\usage{
make_grid(
  x,
  hexagonal = TRUE,
  cell_area,
  clip = FALSE,
  tolerance = NULL,
  grid_boundary = FALSE
)
}
\arguments{
\item{x}{Object of class sf, sfc, sfg or SpatialPolygons}

\item{hexagonal}{logical. If TRUE then the geometry grid will be hexagonal othwewise will be square}

\item{cell_area}{numeric. Grid area in the CRS units of object x.}

\item{clip}{logical. If TRUE, the new grid will be clipped to the x area.
The operation time will be longer the greater the number of vertices in the polygon of the region.
If it is a region with many vertices use the argument **tolerance**.}

\item{tolerance}{numeric. If Reduces the number of vertices in the region polygon.}

\item{grid_boundary}{logical.If TRUE, the "hexagons" or "squares" in the boundaries will be removed.}
}
\description{
Make grid
}
