% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_conefor.R
\name{merge_conefor_mult}
\alias{merge_conefor_mult}
\title{Multiple join connectivity index (dIIC o dPC, and the "intra",  "dflux", "connector") to core areas}
\usage{
merge_conefor_mult(filename, merge_shape, patterns, id, foldname, write,
  merge_coneforlist, prefix)
}
\arguments{
\item{filename}{character. Direction of node_importances.txt files}

\item{merge_shape}{SpatialPolygons. Core area shapefile. It has to have the same id used to estimate the node importance.txt}

\item{patterns}{character. Prefixes, e.g., c("^AREA_25KM.*txt","^AREA_50KM.*txt","^AREA_100KM.*txt")}

\item{id}{character. Column name with the core area id}

\item{foldname}{character. Folders names, these folders will be created inside the outputdir}

\item{write}{character. Output direction}

\item{merge_coneforlist}{list. A list of four arguments used in merge_conefor() function: list(colnames, dA, var)}

\item{prefix}{character. An initial prefix used to write the shapefile, e.g., "AREA"}
}
\description{
Multiple join connectivity index (dIIC o dPC, and the "intra",  "dflux", "connector") to core areas
}
