% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_class.R
\name{metric_class}
\alias{metric_class}
\title{Generates metric class}
\usage{
metric_class(
  metric = NULL,
  attribute = NULL,
  thintersect = NULL,
  distance_threshold = NULL,
  probability = NULL,
  transboundary = NULL,
  distance = NULL
)
}
\arguments{
\item{metric}{character}

\item{attribute}{character}

\item{thintersect}{numeric}

\item{distance_threshold}{numeric}

\item{probability}{numeric}

\item{transboundary}{numeric}

\item{distance}{list}
}
\description{
Generates metric class
}
\keyword{internal}
