% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_BCentrality.R
\name{MK_BCentrality}
\alias{MK_BCentrality}
\title{Betweenness Centrality metrics}
\usage{
MK_BCentrality(nodes, id, attribute = NULL, area_unit = "ha",
  distance = list(type = "centroid", resistance = NULL),
  metric = c("BC", "BCIIC", "BCPC"), distance_thresholds = NULL,
  probability = NULL, LA = NULL, dA = FALSE, dvars = FALSE,
  write = NULL)
}
\arguments{
\item{nodes}{Object of class sf, sfc, sfg or SpatialPolygons. The shapefile must be in a projected coordinate system.}

\item{id}{character. Column name with the nodes id. If NULL, then a new temporal id will be generated.}

\item{attribute}{character. Column name with the nodes attribute. If NULL, then the patch area (ha) will be estimated and used as the attribute.}

\item{area_unit}{character. If attribute is NULL you can set an area unit, udunits2 package compatible unit (e.g., "km2", "cm2", "ha"). Default equal to square meters "ha".}

\item{distance}{list. Distance parameters. For example: type, resistance,or tolerance. For "type" choose one of the distances: "centroid" (faster), "edge",
"least-cost" or "commute-time". If the type is equal to "least-cost" or "commute-time", then you have to use the "resistance" argument.
 To See more arguments consult the help function of distancefile().}

\item{metric}{character. Choose a Betweenness Centrality Metric: "BC" or "BCIIC" considering topologycal distances or "BCPC" considering maximum product probabilities.}

\item{distance_thresholds}{numeric. Distance or distances thresholds (meters) to establish connections. For example, one distance: distance_threshold = 30000; two or more specific distances:
distance_thresholds = c(30000, 50000); sequence distances: distance_thresholds = seq(10000,100000, 10000).}

\item{probability}{numeric. Connection probability to the selected distance threshold, e.g., 0.5 that is 50 percentage of probability connection. Use in case of selecting the "BPC" metric.}

\item{LA}{numeric. Maximum landscape attribute (attribute unit, if attribute is NULL then unit is equal to ha).}

\item{dA}{logical. If TRUE, then the delta attribute will be added to the node's importance result.}

\item{dvars}{logical. If TRUE, then the absolute variation will be added to the node's importance result.}

\item{write}{character. Write output shapefile, example, "C:/ejemplo.shp".}
}
\description{
Use this function to calculate the BC, BCIIC and BCPC indexes under one or several distance thresholds.
}
\examples{
data("vegetation_patches", package = "Makurhini")
nrow(vegetation_patches) #Number of patches

#One distance threshold
MK_BCentrality(nodes = vegetation_patches, id = "id",
            distance = list(type = "centroid"),
            metric = "BCIIC", LA = NULL,
            distance_thresholds = 30000) #30 km

#Two or more distance thresholds
MK_BCentrality(nodes = vegetation_patches, id = "id", attribute = NULL,
           distance = list(type = "centroid"),
           metric = "BCIIC", LA = NULL,
           distance_thresholds = c(10000, 30000)) #10 and 30 km
}
\references{
Saura, S. and Torne, J. (2012). Conefor 2.6. Universidad Politecnica de Madrid. Available at \url{www.conefor.org}.\cr
 Freeman L.C. (1977). Set of Measures of Centrality Based on Betweenness. Sociometry 40: 35-41.\cr
 Bodin, O. and Saura, S. (2010). Ranking individual habitat patches as connectivity providers: integrating network analysis and patch removal experiments. Ecological Modelling 221: 2393-2405.
}
