% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_dECA.R
\name{MK_dECA}
\alias{MK_dECA}
\title{Estimate the ECA, rECA, dA and dECA.}
\usage{
MK_dECA(
  nodes,
  attribute = NULL,
  area_unit = "m2",
  weighted = FALSE,
  distance = list(type = "centroid", resistance = NULL),
  threshold = NULL,
  metric = "IIC",
  probability = NULL,
  distance_thresholds = NULL,
  LA = NULL,
  plot = FALSE,
  parallel = NULL,
  parallel_mode = 1,
  write = NULL,
  intern = TRUE
)
}
\arguments{
\item{nodes}{\code{list}. List of objects containing nodes (e.g., habitat patches or fragments) of each time to analyze information. Nodes can be of the following classes:\cr
-   \code{Data.frame} with at least two columns: the first for node IDs and the second for attributes.\cr
-   Spatial data of type vector (class \code{sf, SpatVector, SpatialPolygonsDataFrame}). It must be in a projected coordinate system.\cr
-   Raster (class \code{RasterLayer, SpatRaster}). It must be in a projected coordinate system. The values must be integers representing the ID of each habitat patch or node, with non-habitat areas represented by NA values (see \link[raster]{clump} or \link[terra]{patches}).}

\item{attribute}{\code{character} or \code{list}. If \code{NULL} (only applicable when \code{nodes} is of spatial data of vector or raster type) the area of the nodes will be used as the node attribute. The unit of area can be selected using the \code{area_unit} parameter. To use an alternative attribute, consider the class type of the object in the \code{nodes} parameter: \cr
-   If \code{nodes} is \code{list} of spatial vectors or data.frames, specify \bold{the name of the column} containing the attribute for the nodes. The column name must be present in each element of the node list.\cr
-   If \code{nodes} is a \code{list} of raster layers then the list must contain numeric vectors with attributes for each period or element of the list of nodes. The first numeric vector in the list must correspond to the first element of the node list. The length of each vector must be equal to the corresponding number of nodes. If the parameter \bold{weighted} is \code{TRUE} then the numeric vector is multiplied by the area of each node to obtain a weighted habitat index.}

\item{area_unit}{\code{character}. (\emph{optional, default = } \code{"m2"}) \cr. A \code{character} indicating the area units when \code{attribute} is \code{NULL}. Some options are "m2" (the default), "km2", "cm2", or "ha";  See \link[Makurhini]{unit_convert} for details.}

\item{weighted}{\code{logical}. If the \code{nodes} are of raster type, you can weight the estimated area of each node by the attribute. When using this parameter the \code{attribute} parameter, which must be a vector of length equal to the number of nodes, usually has values between 0 and 1.}

\item{distance}{A \code{matrix} or \code{list} to establish the distance between each pair of nodes. Distance between nodes may be Euclidean distances (straight-line distance) or effective distances (cost distances) by considering the landscape resistance to the species movements.\cr
- If it is a \code{matrix}, then the number of columns and rows must be equal to the number of nodes. This distance matrix could be generated by the \link[Makurhini]{distancefile} function.\cr
- If it is a \code{list} of parameters, then it must contain the distance parameters necessary to calculate the distance between nodes. For example, two of the most important parameters: \code{“type”} and \code{“resistance”}. For \code{"type"} choose one  of the distances:  \bold{"centroid" (faster), "edge", "least-cost" or "commute-time"}. If the type is equal to \code{"least-cost"} or \code{"commute-time"}, then you must use the \code{"resistance"} argument. For example: \code{distance(type = "least-cost", resistance = raster_resistance)}. To see more arguments see the \link[Makurhini]{distancefile} function.\cr
- You can place a list with resistances where there must be one resistance for each time/scenario of patches in the \code{nodes} parameter, for example,
 if nodes has a list of two times then you can use two resistances one for time 1 and one for time 2:
 \code{distance(type = "least-cost", resistance = list(resistanceT1, resistanceT2))}.}

\item{threshold}{\code{numeric}. Pairs of nodes with a distance value greater than this threshold will be discarded in the analysis which can speed up processing.}

\item{metric}{A \code{character} indicating the connectivity metric to use: \code{"PC"} (the default and recommended) to calculate the probability of connectivity index, and \code{"IIC"} to calculate the binary integral index of connectivity.}

\item{probability}{A \code{numeric} value indicating the probability that corresponds to the distance specified in the \code{distance_threshold}. For example, if the \code{distance_threshold} is a median dispersal distance, use a probability of 0.5 (50\%). If the \code{distance_threshold} is a maximum dispersal distance, set a probability of 0.05 (5\%) or 0.01 (1\%). Use in case of selecting the \code{"PC"} metric. If \code{probability = NULL}, then a probability of 0.5 will be used.}

\item{distance_thresholds}{A \code{numeric} indicating the dispersal distance or distances (meters) of the considered species. If \code{NULL} then distance is estimated as the median dispersal distance between nodes. Alternatively, the \link[Makurhini]{dispersal_distance} function can be used to estimate the dispersal distance using the species home range. Can be the same length as the \code{distance_thresholds} parameter.}

\item{LA}{\code{numeric}. (\emph{optional, default = } \code{NULL}). The maximum landscape attribute, which is the attribute value that would correspond to a hypothetical habitat patch covering all the landscape with the best possible habitat, in which IIC and PC would be equal to 1. For example, if nodes attribute corresponds to the node area, then LA equals total landscape area. If nodes attribute correspond to a quality-weighted area and the quality factor ranges from 0 to 100, LA will be equal to 100 multiplied by total landscape area. The value of LA does not affect at all the importance of the nodes and is only used to calculate the overall landscape connectivity. If no LA value is entered (default) and  \code{overall = TRUE} or \code{onlyoverall = TRUE}, the function will only calculate the numerator of the global connectivity indices and the equivalent connected ECA or EC index.}

\item{plot}{\code{logical}. Also, you can provide the corresponding year for each period of
time analyzed, e.g., c("2011", "2014", "2017")}

\item{parallel}{(\emph{optional, default =} \code{NULL}).
A \code{numeric} specifying the number of cores to parallelize the index estimation of the PC or IIC index and its deltas.Particularly useful when you have more than 1000 nodes. By default the analyses are not parallelized.}

\item{parallel_mode}{(\emph{optional, default =} \code{1}).
A \code{numeric} indicating the mode of parallelization: Mode \bold{1} (the default option, and recommended for less than 1000 nodes) parallelizes on the connectivity delta estimate, while Mode \bold{2} (recommended for more than 1000 nodes)
parallelizes on the shortest paths between vertices estimate.}

\item{write}{\code{character}. Path and name of the output ".csv" file}

\item{intern}{\code{logical}. Show the progress of the process, default = TRUE. Sometimes the advance process does not reach 100 percent when operations are carried out very quickly.}
}
\value{
Table with:\cr\cr
-  Time: name of the time periods, name of the model or scenario (are taken from the name of the elements of the list of nodes or the plot argument)\cr
-  Max. Landscape attribute: maximum landscape attribute\cr
-  Habitat area,\cr
-  Distance threshold: it is usually a dispersal threshold associated with one or many species and it is set by the user.\cr
-  ECA: Equivalent Connected Area or Equivalent Connectivity
-  Normalized_ECA (% of LA): relative connectivity (percentage)\cr
-  Normalized_ECA (% of habitat area): relative connectivity (percentage)\cr
-  dA: delta Area between times (percentage)\cr
-  dECA: delta ECA between times (percentage)\cr
-  rECA: relativized ECA (dECA/dA). According to Liang et al. (2021) "an rECA value greater than 1 indicates that habitat changes result in a
disproportionately large change in habitat connectivity, while a value lower than 1 indicates connectivity
changes due to random habitat changes (Saura et al. 2011; Dilts et al. 2016)".\cr
-  dA/dECA comparisons: comparisons between dA and dECA\cr
-  Type of change: Type of change using the dECAfun() and the difference between dA and dECA.\cr
}
\description{
Equivalent Connected Area (ECA; if the area is used as attribute) or Equivalent Connectivity index (EC)
}
\examples{
\dontrun{
library(Makurhini)
library(sf)

data("list_forest_patches", package = "Makurhini")
data("study_area", package = "Makurhini")
class(list_forest_patches)

Max_attribute <- unit_convert(st_area(study_area), "m2", "ha")

dECA_test <- MK_dECA(nodes= list_forest_patches, attribute = NULL, area_unit = "ha",
                  distance = list(type= "centroid"), metric = "PC",
                  probability = 0.05, distance_thresholds = 5000,
                  LA = Max_attribute, plot= c("1993", "2003", "2007", "2011"),
                  intern = TRUE)
dECA_test

}
}
\references{
\url{www.conefor.org}\cr\cr
- Saura, S., Estreguil, C., Mouton, C., & Rodríguez-Freire, M. (2011). Network analysis to assess landscape connectivity trends: Application to European forests (1990-2000). Ecological Indicators, 11(2), 407–416.
https://doi.org/10.1016/j.ecolind.2010.06.011 \cr
 Herrera, L. P., Sabatino, M. C., Jaimes, F. R., & Saura, S. (2017). Landscape connectivity and the role of small habitat patches as stepping stones: an assessment of the grassland biome in South America. Biodiversity and Conservation, 26(14), 3465–3479.
https://doi.org/10.1007/s10531-017-1416-7\cr
- Liang, J., Ding, Z., Jiang, Z., Yang, X., Xiao, R., Singh, P. B., ... & Hu, H. (2021). Climate change, habitat connectivity, and conservation gaps: a case study of four ungulate species endemic to the Tibetan Plateau. Landscape Ecology, 36(4), 1071-1087.\cr
- Dilts TE, Weisberg PJ, Leitner P, Matocq MD, Inman RD, Nussear KE, Esque TC (2016) Multi-scale connectivity and graph theory highlight critical areas for conservation under climate change. Ecol Appl 26:1223–1237
}
