% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_dECA_grid.R
\name{MK_dECA_grid}
\alias{MK_dECA_grid}
\title{dA and dECA in a regular grid}
\usage{
MK_dECA_grid(
  nodes,
  area_unit = "m2",
  region = NULL,
  grid_param = list(grid_pol = NULL, grid_id = NULL, hexagonal = TRUE, cellsize = NULL,
    grid_boundary = FALSE, clip = FALSE, tolerance = NULL),
  distance = list(type = "centroid"),
  metric = "IIC",
  distance_threshold = NULL,
  probability = NULL,
  intern = TRUE,
  parallel = NULL
)
}
\arguments{
\item{nodes}{list of objects class sf, SpatialPolygonsDataFrame. Nodes of each time to analyze.
The shapefiles must be in a projected coordinate system.}

\item{area_unit}{character. Attribute area units. You can set an area unit, "Makurhini::unit_covert()" compatible unit
("m2", "Dam2, "km2", "ha", "inch2", "foot2", "yard2", "mile2"). Default equal to hectares "ha".}

\item{region}{object of class sf, sfc, sfg or SpatialPolygons. Region shapefile, the shapefile must be
in a projected coordinate system.}

\item{grid_param}{list. Parameters of the grid shapefile, see \link[Makurhini]{get_grid}.Just omit the parameter 'region'. Example,
list(grid_pol = NULL, hexagonal = TRUE, grid_id = NULL, cellsize = unit_convert(1000, "km2", "m2"),
grid_boundary = FALSE, clip = FALSE, tolerance = NULL).}

\item{distance}{list. See \link[Makurhini]{distancefile}. Example, list(type= "centroid", resistance = NULL).}

\item{metric}{character. Choose a connectivity metric: "IIC" considering topologycal distances or "PC"
considering maximum product probabilities.}

\item{distance_threshold}{numeric. Distance threshold to establish connections (crs units, usually meters).}

\item{probability}{numeric. Probability of direct dispersal between nodes, Default, 0.5,
 that is 50 percentage of probability connection. If probability = NULL, then it will be the inverse of the mean dispersal distance
for the species (1/α; Hanski and Ovaskainen 2000).}

\item{intern}{logical. Show the progress of the process, default = TRUE. Sometimes the advance process does not reach 100 percent when operations are carried out very quickly.}

\item{parallel}{numeric. Specify the number of cores to use for parallel processing, default = NULL. Parallelize the function using furrr package and multiprocess
plan when there are more than ONE transboundary.}
}
\description{
Use the function to compute the Protected Connected (ProtConn), EC, PC or IIC indexes in a regular grid.
}
\examples{
\dontrun{
library(Makurhini)
library(sf)

# Four times (T1.2, T2.3, T3.4)
data("list_forest_patches", package = "Makurhini")
data("study_area", package = "Makurhini")
class(list_forest_patches)

hexagons_dECA <- MK_dECA_grid(nodes = list_forest_patches,
                              region = study_area,
                              area_unit = "ha",
                              metric = "IIC",
                              grid_param = list(hexagonal = TRUE,
                                                cellsize = unit_convert(100, "km2", "m2")),
                              distance_threshold = 3000,
                              probability = 0.5,
                              distance = list(type = "centroid"),
                              intern = TRUE,
                              parallel = NULL)
hexagons_dECA
plot(hexagons_dECA["T3.4.dECA"], breaks = "quantile")
plot(hexagons_dECA["T3.4.Type.Change"], key.pos = 1)
}
}
\references{
Matt Strimas-Mackey. \url{http://strimas.com/spatial/hexagonal-grids/}.\cr
Saura, S., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2017). Protected areas in the
world's ecoregions: How well connected are they? Ecological Indicators, 76, 144–158.
Saura, S. & Torne, J. (2012). Conefor 2.6. Universidad Politecnica de Madrid. Available
 at \url{www.conefor.org}.\cr
Pascual-Hortal, L. & Saura, S. (2006). Comparison and development of new graph-based landscape
 connectivity indices: towards the priorization of habitat patches and corridors for conservation.
 Landscape Ecology, 21(7): 959-967.\cr
Saura, S. & Pascual-Hortal, L. (2007). A new habitat availability index to integrate connectivity
in landscape conservation planning: comparison with existing indices and application to a case study.
Landscape and Urban Planning, 83(2-3): 91-103.
}
