% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_dPCIIC_samc.R
\name{MK_dPC_SAMC}
\alias{MK_dPC_SAMC}
\title{Estimate the probability of connectivity (PC) and prioritize nodes using Spatial Absorbing Markov Chains}
\usage{
MK_dPC_SAMC(
  nodes,
  keep = NULL,
  attribute = NULL,
  weighted = FALSE,
  LA = NULL,
  area_unit = "m2",
  restoration = NULL,
  onlyrestor = FALSE,
  disp_kernel = list(resistance = NULL, CostFun = NULL, absorption = NULL, dir = 8,
    SettlFunc = sum, pvertices = NULL, nvertices = 10, systematic = TRUE),
  probability = NULL,
  dispersal_dist = NULL,
  overall = FALSE,
  onlyoverall = FALSE,
  parallel = NULL,
  parallel_mode = 1,
  write = NULL,
  id_sel = NULL,
  intern = TRUE
)
}
\arguments{
\item{nodes}{Object containing nodes (e.g., habitat patches or fragments) information. It can be of the following classes:\cr
-   \code{Data.frame} with at least two columns: the first for node IDs and the second for attributes. If the `restoration` argument is used, the data frame must include a third column for restoration values.\cr
-   Spatial data of type vector (class \code{sf, SpatVector, SpatialPolygonsDataFrame}). It must be in a projected coordinate system.\cr
-   Raster (class \code{RasterLayer, SpatRaster}). It must be in a projected coordinate system. The values must be integers representing the ID of each habitat patch or node, with non-habitat areas represented by NA values (see \link[raster]{clump} or \link[terra]{patches}).}

\item{keep}{numeric. Use this option to simplify the nodes geometry and reduce the number of vertices. The value can range from 0 to 1 and is the proportion of points to retain (default equal to NULL). The higher the value, the higher the speed but less precision.}

\item{attribute}{\code{Character} or \code{vector}. If \code{NULL} (only applicable when \code{nodes} is of spatial data of vector or raster type) the area of the nodes will be used as the node attribute. The unit of area can be selected using the \code{area_unit} parameter. To use an alternative attribute, consider the class type of the object in the \code{nodes} parameter: \cr
-   If \code{nodes} is a spatial vector or data.frame, specify \bold{the name of the column} containing the attribute for the nodes. \cr
-   If \code{nodes} is a raster layer then it must be a numeric vector with the node's attribute. The length of the vector must be equal to the number of nodes. If the parameter \bold{weighted} is \code{TRUE} then the numeric vector is multiplied by the area of each node to obtain a weighted habitat index.}

\item{weighted}{\code{logical}. If the \code{nodes} are of raster type, you can weight the estimated area of each node by the attribute. When using this parameter the \code{attribute} parameter, which must be a vector of length equal to the number of nodes, usually has values between 0 and 1.}

\item{LA}{\code{numeric}. (\emph{optional, default = } \code{NULL}). The maximum landscape attribute, which is the attribute value that would correspond to a hypothetical habitat patch covering all the landscape with the best possible habitat, in which PC would be equal to 1. For example, if nodes attribute corresponds to the node area, then LA equals total landscape area. If nodes attribute correspond to a quality-weighted area and the quality factor ranges from 0 to 100, LA will be equal to 100 multiplied by total landscape area. The value of LA does not affect at all the importance of the nodes and is only used to calculate the overall landscape connectivity. If no LA value is entered (default) and  \code{overall = TRUE} or \code{onlyoverall = TRUE}, the function will only calculate the numerator of the global connectivity indices and the equivalent connected ECA or EC index.}

\item{area_unit}{\code{character}. (\emph{optional, default = } \code{"m2"}) \cr. A \code{character} indicating the area units when \code{attribute} is \code{NULL}. Some options are "m2" (the default), "km2", "cm2", or "ha";  See \link[Makurhini]{unit_convert} for details.}

\item{restoration}{\code{Character or vector}, (\emph{optional}). This parameter specifies the binary restoration value indicating whether each node is existing or hypothetically added for restoration: 1 for existing nodes in the landscape and 0 for nodes to be added to the initial landscape. If \code{NULL} (default), all nodes are considered existing (as if all restoration values were 1). Otherwise: \cr
-   If \code{nodes} is a shapefile (spatial vector) or a data.frame, specify the name of the column containing the restoration values. If `nodes` is a data.frame, it must have three columns: the first for node IDs, the second for attributes, and the third for restoration values.\cr
-   If \code{nodes} is a raster layer, provide a numeric vector with restoration values for each node in the raster.}

\item{onlyrestor}{\code{logical}. If \code{TRUE}, then only restoration metric will be calculated.}

\item{disp_kernel}{A \code{list} with the parameters used to estimate settlement probability.
Expected components:
\describe{
  \item{\code{resistance}}{Resistance raster (e.g., \code{terra::SpatRaster})
  where higher values indicate greater movement cost. The resistance raster must be a closed matrix with no isolated pixels.}

  \item{\code{CostFun}}{Function to compute the cost of moving between cells
  from local resistance. Default: \code{function(x) 1/mean(x)}.}

  \item{\code{absorption}}{raster with absorption rate for absorbing states. By default,
  it is estimated from the species’ mean dispersal distance and the cell size
  of the resistance raster (Fletcher et al., 2023). If raster, then it must be a closed matrix with no isolated pixels.}

  \item{\code{dir}}{Neighborhood (directions) for the probability graph.
  Options: \code{4}, \code{8} (Moore, default), or \code{16}.}

  \item{\code{SettlFunc}}{Function that aggregates (combines) settlement
  probabilities among polygon vertices (edges) for each patch. Default:
  \code{mean}.}

  \item{\code{pvertices}}{Proportion of boundary vertices per patch to include
  in the analysis (range \code{(0, 1]}). Default: \code{0.1}.}

   \item{\code{nvertices}}{Number of boundary vertices per patch to include
  in the analysis. Default: \code{0.1}.}

   \item{\code{systematic}}{If TRUE, vertices are selected systematically, with each selected vertex separated by approximately the same number of intermediate vertices. When FALSE, the function selects vertices at random. Default: \code{TRUE}.}
}}

\item{probability}{A \code{numeric} value indicating the probability that corresponds to the distance specified in the \code{dispersal_dist}. For example, if the \code{dispersal_dist} is a median dispersal distance, use a probability of 0.5 (50\%). If the \code{dispersal_dist} is a maximum dispersal distance, set a probability of 0.05 (5\%) or 0.01 (1\%). Use in case of selecting the \code{"PC"} metric. If \code{probability = NULL}, then a probability of 0.5 will be used.}

\item{dispersal_dist}{A \code{numeric} indicating the dispersal distance (meters) of the considered species. If \code{NULL} then distance is estimated as the median dispersal distance between nodes. Alternatively, the \link[Makurhini]{dispersal_distance} function can be used to estimate the dispersal distance using the species home range.}

\item{overall}{\code{logical}. If \code{TRUE}, then the overall metrics will be added to the result which is transformed into a list. Default equal to FALSE}

\item{onlyoverall}{\code{logical}. If \code{TRUE}, then only overall metrics will be calculated.}

\item{parallel}{(\emph{optional, default =} \code{NULL}).
A \code{numeric} specifying the number of cores to parallelize the index estimation of the PC index and its deltas.Particularly useful when you have more than 1000 nodes. By default the analyses are not parallelized.}

\item{parallel_mode}{(\emph{optional, default =} \code{1}).
A \code{numeric} indicating the mode of parallelization: Mode \bold{1} (the default option, and recommended for less than 1000 nodes) parallelizes on the connectivity delta estimate, while Mode \bold{2} (recommended for more than 1000 nodes)
parallelizes on the shortest paths between vertices estimate.}

\item{write}{\code{Character} indicating the path and initial prefix of the objects to save, for example, \code{"C:/example/test_PC_"}. By default, nothing is saved. The saved objects are: \cr
-   The importance of each node or habitat patch. The format of the output file depends on the class of \code{nodes} (\code{shapefile, raster, or table}).\cr
-   Overall landscape connectivity table (if the \code{overall} argument is \code{TRUE}).}

\item{id_sel}{Internal use only, not for users.}

\item{intern}{\code{logical}. Show the progress of the process, \code{default = TRUE}. Sometimes the advance process does not reach 100 percent when operations are carried out very quickly.}
}
\value{
-   If you add \bold{\code{overall = TRUE}}, then a list containing the \code{sf} class object with the importance values of the nodes and a \code{data.frame} with the overall connectivity values (num, EC, PC) will be returned.\cr
-   If you use the \bold{\code{restoration}} parameter then an extra column will be returned to the \code{sf} object with the node importance values, unless you use the \code{onlyrestor} argument (i.e., equal to \code{TRUE}) only the restoration metric is estimated.\cr
}
\description{
Computes overall landscape connectivity and the importance (contribution) of
each node (habitat patch) to that connectivity using the Probability of
Connectivity (PC) index for a specified dispersal distance. Movement and
settlement are modeled under a Spatial Absorbing Markov Chain (SAMC)
framework.
}
\details{
The SAMC framework treats raster cells as states with three roles: departure, transient
(movement) and absorbing (settlement at patch boundaries). Transition
probabilities are derived from a resistance surface (and a chosen neighborhood
structure), while absorption rates reflect species’ dispersal and patch
properties. Multi-step transitions yield probabilities of reaching (and
settling in) patches across the landscape. These probabilities are then used
to compute the \strong{overall connectivity} (PC) and the \strong{node
importance} (each patch’s contribution to PC, e.g., via ΔPC). The SAMC
computations are implemented via the \pkg{samc} R package (Marx et al., 2020).
}
\note{
Sometimes the advance process does not reach 100 percent when operations are carried out very quickly.
}
\examples{
\dontrun{
library(Makurhini)
library(sf)
library(raster)
data("habitat_nodes", package = "Makurhini")
nrow(habitat_nodes) # Number of patches
data("resistance_matrix", package = "Makurhini")


#Example
hn <- habitat_nodes[350:404,]
hn$Id <- 1:nrow(hn)
plot(hn)
mask_hn <- st_as_sfc(st_bbox(hn)) |> st_as_sf() |> st_buffer(5000)
#Aggregate for this example
resist <- aggregate(resistance_matrix, 4)
resist <- crop(resistance_matrix, mask_hn)
plot(resist)

#Two distance threshold,
PCsamc <- MK_dPC_SAMC(nodes = hn,
                      keep = 0.5,
                      attribute = NULL,
                      area_unit = "m2",
                      disp_kernel = list(resistance = resist,
                                         CostFun = function(x) 1/mean(x),
                                         absorption = NULL, dir = 4,
                                         SettlFunc = sum, nvertices = 10),
                      LA = NULL,
                      overall = TRUE,
                      dispersal_dist = 10000) #10 km
PCsamc$overall_d10000
plot(PCsamc$node_importances_d10000["dPC"], breaks = "jenks")
plot(PCsamc$node_importances_d10000["dPCintra"], breaks = "jenks")
plot(PCsamc$node_importances_d10000["dPCflux"], breaks = "jenks")
plot(PCsamc$node_importances_d10000["dPCconnector"], breaks = "jenks")

}
}
\references{
-   Saura, S. & Torné, J. 2012. Conefor 2.6 user manual (May 2012). Universidad Politécnica de Madrid. Available at \url{www.conefor.org}.\cr
-   Saura, S. & Pascual-Hortal, L. 2007. A new habitat availability index to integrate connectivity in landscape conservation planning: comparison with existing indices and application to a case study. Landscape and Urban Planning 83 (2-3): 91-103.\cr
-   Saura, S., Bodin, Ö., & Fortin, M.-J. (2014). EDITOR’S CHOICE: Stepping stones are crucial for species’ long-distance dispersal and range expansion through habitat networks. Journal of Applied Ecology, 51(1), 171-182.\cr
-   Hanski, I. and Ovaskainen, O. 2000. The metapopulation capacity of a fragmented landscape. Nature 404: 755–758.\cr
-   Marx, A. J., Wang, C., Sefair, J. A., Acevedo, M. A., & Fletcher Jr., R. J. (2020). samc: An R package for connectivity modeling with spatial absorbing Markov chains. Ecography, 43(4), 518-527.\cr
-   Fletcher, R. J., Iezzi, M. E., Guralnick, R., Marx, A. J., Ryan, S. J., & Valle, D. (2023). A framework for linking dispersal biology to connectivity across landscapes. Landscape Ecology, 38(10), 2487-2500. https://doi.org/10.1007/s10980-023-01741-8
}
