% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_ProtConn.R
\name{MK_ProtConn}
\alias{MK_ProtConn}
\title{Protected Connected (ProtConn)}
\usage{
MK_ProtConn(
  nodes,
  region,
  thintersect = NULL,
  attribute = "Intersected area",
  area_unit = "ha",
  res_attribute = 150,
  distance = list(type = "centroid", resistance = NULL),
  distance_thresholds,
  probability,
  transboundary = NULL,
  LA = NULL,
  keep = 1,
  plot = FALSE,
  dPC = FALSE,
  write = NULL,
  SAGA = FALSE,
  intern = TRUE
)
}
\arguments{
\item{nodes}{object of class sf, sfc, sfg or SpatialPolygons. Example, Protected areas shapefile.}

\item{region}{object of class sf, sfc, sfg or SpatialPolygons. Region shapefile.}

\item{thintersect}{numeric. Threshold of intersection in percentage allowed to select or not a target geometry.
Example, if thintersect is equal to 90 then a node will be selected only if the intersection between the node and
 the region is >= 90 percentage. If NULL, thintersect will be 0 (default)}

\item{attribute}{character. Select the nodes attribute: "Intersected area" = Intersected Protected areas; or
another specific column name with the nodes attribute, ideally this attribute mus be an area-weighted index,
 otherwise the interpretation of the protconn index may change.}

\item{area_unit}{character. Attribute area units. You can set an area unit, "Makurhini::unit_covert()" compatible unit ("m2", "Dam2, "km2", "ha", "inch2", "foot2", "yard2", "mile2"). Default equal to hectares "ha".}

\item{res_attribute}{numeric. If the attribute is no equal to "Area" or "Intersected area" then nodes will be converted to raster to extract values in one  process step, you can set the raster resolution, default = 150.}

\item{distance}{list. See distancefile(). E.g.: list(type= "centroid", resistance = NULL).}

\item{distance_thresholds}{numeric. Distance or distances thresholds to establish connections (meters). For example, one distance: distance_threshold = 30000; two or more specific distances:
distance_thresholds = c(30000, 50000); sequence distances: distance_thresholds = seq(10000,100000, 10000).}

\item{probability}{numeric. Connection probability to the selected distance threshold, e.g., 0.5 that is 50 percentage of probability connection.
If probability = NULL, then it will be the inverse of the mean dispersal distance for the species (1/α; Hanski and Ovaskainen 2000).}

\item{transboundary}{numeric. Buffer to select polygones in a second round, their attribute value = 0, see Saura et al. 2017. You can set one transboundary value or one per each threshold distance.}

\item{LA}{numeric. Maximum Landscape Attribute.}

\item{keep}{numeric. Simplification of the region's geometry to speed the select by location ("Makurhini::MK_selectbyloc").
The value can range from 0 to 1 and is the proportion of points to retain (default 1). The lower the value,
 the higher the speed but the greater uncertainty.}

\item{plot}{logical. Plot the main ProtConn indicators and fractions, default = FALSE.}

\item{dPC}{logical. if TRUE dPC indexes are merged with the nodes shapefile, default = FALSE.}

\item{write}{character. Output folder including the output file name without extension, e.g., "C:/ProtConn/Protfiles".}

\item{SAGA}{Logical. Optimize the large process using SAGA GIS and RSAGA package (see, \url{https://github.com/r-spatial/RSAGA}).}

\item{intern}{logical. Show the progress of the process, default = TRUE.}
}
\value{
Table with the following ProtConn values: ECA, Prot, ProtConn, ProtUnconn, RelConn, ProtConn[design], ProtConn[bound], ProtConn[Prot], ProtConn[Within],
 ProtConn[Contig], ProtConn[Trans], ProtConn[Unprot], ProtConn[Within][land], ProtConn[Contig][land],
 ProtConn[Unprot][land], ProtConn[Trans][land] \cr
\cr
*If plot is not NULL a list is returned with the ProtConn table and a plots.
}
\description{
Estimate Protected Connected (ProtConn) indicator and fractions for one region.
}
\examples{
\dontrun{
library(Makurhini)
library(raster)
data("Protected_areas", package = "Makurhini")
#plot(Protected_areas, col="green")

data("regions", package = "Makurhini")
region <- regions[1,]
plot(region, col="blue")

test <- MK_ProtConn(nodes = Protected_areas, region = region,
                    attribute = "Intersected area", area_unit = "ha",
                    distance = list(type= "centroid"),
                    distance_thresholds = c(50000, 10000),
                    probability = 0.5, transboundary = 50000,
                    LA = NULL, plot = TRUE, dPC = FALSE,
                    write = NULL, SAGA = FALSE, intern = TRUE)
test$d50000
test$d10000
}
}
\references{
Saura, S., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2017). Protected areas in the world’s ecoregions: How well connected are they? Ecological Indicators, 76, 144–158.
Saura, S., Bertzky, B., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2018). Protected area connectivity: Shortfalls in global targets and country-level priorities. Biological Conservation, 219(October 2017), 53–67.
}
