% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_dPCIIC.R
\name{nodes_dPC_res}
\alias{nodes_dPC_res}
\title{Estimate restoration using PC}
\usage{
nodes_dPC_res(
  idres = NULL,
  graph_n = NULL,
  attrib_n = NULL,
  num = NULL,
  G1 = 1000,
  min_nodes = 2000,
  parallel = NULL,
  parallel_mode = 0,
  intern = FALSE
)
}
\arguments{
\item{idres}{integer. Id patches = 0}

\item{graph_n}{Graph from cppRouting package}

\item{attrib_n}{numeric. Nodes attribute}

\item{num}{numeric. numPC ID patches = 1}

\item{G1}{numeric}

\item{min_nodes}{numeric}

\item{parallel}{numeric}

\item{parallel_mode}{numeric}

\item{intern}{logical}
}
\description{
Estimate restoration using PC
}
\keyword{internal}
