% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodesfile.R
\name{nodesfile}
\alias{nodesfile}
\title{Creation of nodefile to CONEFOR command line}
\usage{
nodesfile(nodes, id, attribute = NULL, area_unit = "m2",
  restauration = NULL, multiple = NULL, prefix = NULL,
  write = NULL)
}
\arguments{
\item{nodes}{sp or sf object. Nodes shapefile (lines, points or polygons).If area_unit is used then the shapefile must be in a projected coordinate system.}

\item{id}{character. Column name with the node ID in the shapefile data table.}

\item{attribute}{character. Column name with the attribute in the data table selected for the nodes. If NULL the node area will be used as a node attribute, the unit area can be selected using the "area_unit" argument.}

\item{area_unit}{character. If attribute is NULL you can set an area unit, "Makurhini::unit_covert()" compatible unit ("m2", "Dam2, "km2", "ha", "inch2", "foot2", "yard2", "mile2"). Default equal to "m2".}

\item{restauration}{character. Name of the column with restauration value. Binary values (0,1), where 1 = existing nodes in the landscape, and 0 = a new node to add to the initial landscape (restored).}

\item{multiple}{character. Name of the column with the regions names. Use in case of processing nodes of several independent sites at the same time.}

\item{prefix}{character. Initial prefix, use in case of processing several sites at the same time in CONEFOR command line.}

\item{write}{character. Output folder path if you use the name option, otherwise, place the output path, with the name and extension ".txt"}
}
\value{
nodo file in .txt format
}
\description{
Generates a table with the Core ID and the their attribute value (e.g. area in km²), using a shapefile with these values. The table is the nodefile input for CONEFOR.
}
