% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over_poly.R
\name{over_poly}
\alias{over_poly}
\title{over sf}
\usage{
over_poly(x, y, geometry = FALSE)
}
\arguments{
\item{x}{Object of class sf, sfc, sfg or SpatialPolygons}

\item{y}{Object of class sf, sfc, sfg or SpatialPolygons}

\item{geometry}{logical}
}
\description{
over sf
}
