% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIntercsect.R
\name{PIntercsect}
\alias{PIntercsect}
\title{intersect polygons}
\usage{
PIntercsect(p1, p2)
}
\arguments{
\item{p1}{object of class sf, sfc, sfg or SpatialPolygons. Polygons 1 (polygons to be parallelized)}

\item{p2}{object of class sf, sfc, sfg or SpatialPolygons. Polygons 2}
}
\description{
intersect polygons
}
\keyword{internal}
