% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotprotconn.R
\name{plotprotconn}
\alias{plotprotconn}
\title{Protected Connected (ProtConn)}
\usage{
plotprotconn(DataProtconn, d)
}
\arguments{
\item{DataProtconn}{data.frame}

\item{d}{distance threshold}
}
\description{
Estimate Protected Connected (ProtConn) indicator and fractions for one region.
}
\keyword{internal}
