% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pol_dissolve.R
\name{pol_dissolve}
\alias{pol_dissolve}
\title{Dissolve polygones sf or SpatialPolygonsDataFrame}
\usage{
pol_dissolve(polygone, attribute)
}
\arguments{
\item{polygone}{object of class sf, sfc, sfg or SpatialPolygons}

\item{attribute}{character. Column name with the dissolve attribute}
}
\description{
Dissolve polygones sf or SpatialPolygonsDataFrame
}
\keyword{internal}
