% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protconn_dist.R
\name{protconn_dist}
\alias{protconn_dist}
\title{Estimates nodes distances to ProtConn}
\usage{
protconn_dist(x, id, y, r = NULL, resistance = NULL, resist.units = FALSE)
}
\arguments{
\item{x}{object of class sf, sfc, sfg or SpatialPolygons.}

\item{id}{character.}

\item{y}{list. Distance list.}

\item{r}{object of class sf, sfc, sfg or SpatialPolygons.}

\item{resistance}{raster}

\item{resist.units}{logical. Transform cost units to geographic units by multiplying the
cost by the resolution of the raster}
}
\description{
Estimates nodes distances to ProtConn
}
