% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_ProtConn.R
\name{ProtConn_Estimation}
\alias{ProtConn_Estimation}
\title{Protconn estimation}
\usage{
ProtConn_Estimation(
  base_param3 = NULL,
  base_param1 = NULL,
  nodes.delta = NULL,
  n = NULL,
  delta = FALSE,
  transboundary_type = NULL,
  transboundary = NULL,
  protconn_bound = FALSE,
  plot = TRUE,
  LA = NULL,
  intern = TRUE,
  write = NULL
)
}
\arguments{
\item{base_param3}{list}

\item{base_param1}{list}

\item{nodes.delta}{sf}

\item{n}{numeric}

\item{delta}{logical}

\item{transboundary_type}{character}

\item{transboundary}{numeric}

\item{protconn_bound}{logical}

\item{plot}{logical}

\item{LA}{numeric}

\item{intern}{logical}

\item{write}{character}
}
\description{
Protconn estimation
}
\keyword{internal}
