% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_ProtConn_raster.R
\name{ProtConn_Estimation_raster}
\alias{ProtConn_Estimation_raster}
\title{Protconn estimation raster}
\usage{
ProtConn_Estimation_raster(
  patches = NULL,
  n = NULL,
  distance = NULL,
  intern = TRUE,
  distance_thresholds = NULL,
  probability = NULL,
  delta = FALSE,
  region = NULL,
  resist = NULL,
  plot = TRUE,
  transboundary = NULL,
  protconn_bound = NULL,
  LA = NULL,
  works = NULL,
  write = NULL
)
}
\arguments{
\item{patches}{list}

\item{n}{numeric}

\item{distance}{list}

\item{intern}{logical}

\item{distance_thresholds}{numeric}

\item{probability}{numeric}

\item{delta}{logical}

\item{plot}{logical}

\item{LA}{numeric}

\item{works}{numeric}

\item{write}{character}
}
\description{
Protconn estimation raster
}
\keyword{internal}
