% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Protconn_nodes.R
\name{Protconn_nodes}
\alias{Protconn_nodes}
\title{Generates ProtConn nodes files}
\usage{
Protconn_nodes(
  x,
  y,
  buff = NULL,
  xsimplify = FALSE,
  metrunit = "ha",
  protconn = TRUE,
  protconn_bound = FALSE
)
}
\arguments{
\item{x}{object of class sf, sfc, sfg or SpatialPolygons}

\item{y}{object of class sf, sfc, sfg or SpatialPolygons}

\item{buff}{numeric}

\item{xsimplify}{logical}

\item{metrunit}{character}

\item{protconn}{logical. If FALSE then only PC and EC}

\item{protconn_bound}{logical}
}
\description{
Generates ProtConn nodes files
}
