% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtConnCLa.R
\name{ProtConnCLa}
\alias{ProtConnCLa}
\title{Protected Connected (ProtConn)}
\usage{
ProtConnCLa(nodes, region, thintersect = NULL,
  attribute = "Intersected area", res_attribute = 150,
  distance = list(type = "centroid", resistance = NULL, tolerance = NULL,
  geometry_out = NULL, ...), distance_thresholds, probability,
  transboundary = NULL, LA = NULL, plot = FALSE, dPC = FALSE,
  write = NULL, SAGA = FALSE, intern = TRUE)
}
\arguments{
\item{nodes}{object of class sf, sfc, sfg or SpatialPolygons. Protected areas (PAs) shapefile.}

\item{region}{object of class sf, sfc, sfg or SpatialPolygons. Ecoregion shapefile.}

\item{thintersect}{numeric. Threshold of intersection in percentage allowed to select or not a target geometry. Example, if equal to 90 then a nodes will be selected only if the intersection between the node and the region is >= 90 percentage. If NULL, then the nodes will be selected withou}

\item{attribute}{character. Select the nodes attribute: "Area" = Complete Protected areas; "Intersected area" = Intersected Protected areas; or another specific column name with the nodes attribute, ideally this attribute mus be an area-weighted index, otherwise the interpretation of the protconn index may change.}

\item{res_attribute}{numeric. If the attribute is no equal to "Area" or "Intersected area" then nodes will be converted to raster to extract values in one  process step, you can set the raster resolution, default = 150.}

\item{distance}{list.See distancefile(). E.g.: list(type= "centroid", resistance = NULL).}

\item{probability}{numeric. Connection probability to the selected distance threshold, e.g., 0.5 that is 50 percentage of probability connection. Use in case of selecting the "PC" metric.}

\item{transboundary}{numeric. Buffer to select polygones in a second round, their attribute value = 0, see Saura et al. 2017. You can set one transboundary value or one per each threshold distance.}

\item{LA}{numeric. Maximum Landscape Attribute.}

\item{plot}{logical. Plot the main ProtConn indicators and fractions, default = FALSE.}

\item{dPC}{logical. if TRUE dPC indexes are merged with the nodes shapefile, default = FALSE.}

\item{write}{character. Output folder including the output file name without extension, e.g., "C:/ProtConn/Protfiles".}

\item{intern}{logical. Show the progress of the process, default = TRUE.}

\item{distance_threshold}{numeric. Distance or distances thresholds to establish connections. For example, one distance: distance_threshold = 30000; two or more specific distances:
distance_threshold = c(30000, 50000); sequence distances: distance_threshold = seq(10000,100000, 10000).}
}
\value{
Table with the following ProtConn values: ECA, Prot, ProtConn, ProtUnconn, RelConn, ProtConn[design], ProtConn[bound], ProtConn[Prot], ProtConn[Within],
ProtConn[Contig], ProtConn[Trans], ProtConn[Unprot], ProtConn[Within][land], ProtConn[Contig][land],
ProtConn[Unprot][land], ProtConn[Trans][land] \cr
\cr
*If plot is not NULL a list is returned with the ProtConn table and a plots.
}
\description{
Use the CONEFOR command line to estimate Protected Connected (ProtConn) indexes.
}
\references{
Saura, S., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2017). Protected areas in the world’s ecoregions: How well connected are they? Ecological Indicators, 76, 144–158.
Saura, S., Bertzky, B., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2018). Protected area connectivity: Shortfalls in global targets and country-level priorities. Biological Conservation, 219(October 2017), 53–67.
}
