% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtConnStat.R
\name{ProtConnStat}
\alias{ProtConnStat}
\title{Multiple ProtConn statistic}
\usage{
ProtConnStat(ProtConn, ci = "all", nr = 500)
}
\arguments{
\item{ProtConn}{ProtConn table (see, ProtConnCLaMult())}

\item{ci}{character. A character vector representing the type of confidence intervals that will be estimated. The value should be any subset of the values c("norm","basic", "stud", "perc", "bca") or "all" which will compute all five types of intervals (see, boot::boot.ci())}

\item{nr}{numeric. The number of bootstrap replicates}
}
\description{
Multiple ProtConn statistic
}
\references{
Carpenter, J., & Bithell, J. (2000). Bootstrap con " dence intervals : when , which , what ? A practical guide for medical statisticians. Statist. Med., 19, 1141–1164.
}
