% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Protected_areas_raster.R
\docType{data}
\name{Protected_areas_raster}
\alias{Protected_areas_raster}
\title{Protected areas raster for analysis}
\format{
An object of class \code{RasterLayer} of dimension 855 x 684 x 1.
}
\usage{
Protected_areas_raster
}
\description{
raster \code{raster} with 706 polygons of Protected Areas in Mexico.
}
\references{
UNEP-WCMC (2019). Protected Area Profile for Mexico from the World Database of Protected Areas, May 2019.
Available at: www.protectedplanet.net
}
\keyword{datasets}
