% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_clump_points.R
\name{rast_clump_points}
\alias{rast_clump_points}
\title{Get centroid clumb raster}
\usage{
rast_clump_points(x, parallel = NULL, centroid_geometry = FALSE)
}
\arguments{
\item{x}{raster with unique ids}

\item{parallel}{numeric. Specify the number of cores to use for parallel processing}

\item{centroid_geometry}{logical. Return geometry (TRUE) or data.frame(FALSE)}
}
\description{
Get centroid clumb raster
}
\keyword{internal}
