% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_vegetation_patches.R
\docType{data}
\name{raster_vegetation_patches}
\alias{raster_vegetation_patches}
\title{Cores raster for analysis}
\format{A \code{raster} with 142 features and 1 field, which are:
\describe{
\item{id}{Vegetation patches identifier.}
 }}
\usage{
raster_vegetation_patches
}
\description{
A RasterLayer with 142 Vegetation patches with less alteration or degradation that
 is for testing and running of MK_dPCIIC examples.
}
\references{
INEGI. (2013). Conjunto de datos vectoriales de uso del suelo y vegetacion, serie V (capa union), escala 1:250, 000.
Instituto Nacional de Estadistica y Geografia, Aguascalientes.
}
\keyword{datasets}
