% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{regions}
\alias{regions}
\title{Ecoregions shapefile for analysis}
\format{A \code{SpatialPolygonsDataFrame} with 3 features and 3 fields, which are:
\describe{
\item{OBJECTID}{Ecoregions identifier.}
 }}
\usage{
regions
}
\description{
A Spatial Polygons DataFrame \code{SpatialPolygonsDataFrame} with 3 ecoregions of Mexico.
}
\references{
Dinerstein, E., Olson, D., Joshi, A., Vynne, C., Burgess, N. D., Wikramanayake, E., and Kindt, R. (2017).
An Ecoregion-Based Approach to Protecting Half the Terrestrial Realm. BioScience, 67(6), 534?545. https://doi.org/10.1093/biosci/bix014
https://ecoregions2017.appspot.com
}
\keyword{datasets}
