% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resistance_matrix}
\alias{resistance_matrix}
\title{Landscape resistance to dispersal}
\format{
A \code{raster} with Human footprint values from 0 to 100:
\describe{
\item{values}{Human footprint values from 0 to 100}
 }
}
\usage{
resistance_matrix
}
\description{
A RasterLayer with Human footprint values. The raster was aggregated by a factor of 5 to change its original resolution from 100m to 500m.
}
\references{
Correa Ayram, C. A., Mendoza, M. E., Etter, A., & Pérez Salicrup, D. R. (2017). Anthropogenic impact on habitat connectivity: A multidimensional human footprint index evaluated in a highly biodiverse landscape of Mexico. Ecological Indicators, 72, 895-909. https://doi.org/10.1016/j.ecolind.2016.09.007
}
\keyword{datasets}
