% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_dPCIIC_samc.R
\name{samc_settlement}
\alias{samc_settlement}
\title{Estimate probability of settlement}
\usage{
samc_settlement(
  hn = NULL,
  samc_class = NULL,
  pvertices = NULL,
  nvertices = NULL,
  SettlFunc = NULL,
  sampletype = TRUE,
  simpl = NULL,
  intern = TRUE
)
}
\arguments{
\item{hn}{nodes}

\item{samc_class}{samc object}

\item{SettlFunc}{function}

\item{sampletype}{logical}

\item{simpl}{numeric}

\item{intern}{logical}

\item{vertices}{numeric}
}
\description{
Estimate probability of settlement
}
\keyword{internal}
