% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectbyLoc.R
\name{SelectbyLoc}
\alias{SelectbyLoc}
\title{Select layer by location}
\usage{
SelectbyLoc(target, sourcelyr, id = NULL, selreg = "M1",
  buffer = NULL, thintersect = NULL, transboundary = NULL,
  area_unit = "ha", write_select = NULL, plot = FALSE,
  SAGA = FALSE)
}
\arguments{
\item{target}{object of class sf, sfc, sfg or SpatialPolygons.The selection will be applied to this layer.}

\item{sourcelyr}{object of class sf, sfc, sfg or SpatialPolygons.The features in the target layer will be selected based on their relationship to the features from this layer.}

\item{id}{character. Column name of the Core ID.}

\item{selreg}{character. "M1" select by intersect or "M2" select by intersect considering an umbral of intersection (thintersect parameter).
In both options you can choose a transboundary buffer.}

\item{buffer}{numeric. If "M2" is selected, you can choose a buffer for sourcelyr layer (optional).}

\item{thintersect}{numeric. If "M2" is selected, threshold of intersection in percentage allowed to select or not a target geometry. For example,
if a value of 90 then if the intersection area is equal or greater than 90 percent, the geometry will be selected.}

\item{transboundary}{numeric. Transboundary buffer to select polygones in a second round. See, Saura et al. 2017.
Transboundary field is added where 0 = Transboundary.}

\item{area_unit}{character. If attribute is NULL you can set an area unit, udunits2 package compatible unit (e.g., "km2", "cm2", "ha"). Default equal to hectares "ha".}

\item{write_select}{character. Write shapefile, provide folder direction, and name plus extension of the output shapefile}

\item{plot}{logical. Default = FALSE. 1 = Not transboundary PA, 0 = Transboundary PA}

\item{SAGA}{Logical. Optimize the large process using SAGA GIS and RSAGA package (see, \url{https://github.com/r-spatial/RSAGA}).}
}
\value{
shapefile of selected geometries and intersected areas (default equal to hectares)
}
\description{
Select layer by location
}
\examples{
library(Makurhini)
library(raster)
ruta <- system.file("extdata", "WDPA_May2019_MEX-shapefile-polygons.shp", package = "Makurhini")
x <- shapefile(ruta)
plot(x, col="green")

ruta <- system.file("extdata", "Ecoregions2017.shp", package = "Makurhini")
y <- shapefile(ruta)
#Select one region
y_1 <- y[1,]
plot(y_1, col="blue")

#Select polygones x inside y
selection <- SelectbyLoc(target = x, sourcelyr = y, selreg = "M1")
plot(selection, col="green", add=TRUE)
}
