% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_dPCIIC_samc.R
\name{SR_vertices}
\alias{SR_vertices}
\title{Get systematic or random vertices}
\usage{
SR_vertices(
  poly = NULL,
  k = NULL,
  kp = NULL,
  systematic = TRUE,
  return_sf = TRUE
)
}
\arguments{
\item{poly}{nodes}

\item{k}{numeric Number of vertices}

\item{kp}{numeric Proportion of vertices}

\item{systematic}{logical}

\item{return_sf}{logical}
}
\description{
Get systematic or random vertices
}
\keyword{internal}
