% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_centroid_within_poly.R
\name{st_centroid_within_poly}
\alias{st_centroid_within_poly}
\title{st_centroid_within_poly}
\usage{
st_centroid_within_poly(poly)
}
\arguments{
\item{poly}{object of class sf, sfc, sfg or SpatialPolygons}
}
\description{
Polygon centroid
}
\references{
\url{https://stackoverflow.com/questions/52522872/r-sf-package-centroid-within-polygon}
\url{https://stackoverflow.com/users/3609772/mitch}.
}
