% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation_patches.R
\docType{data}
\name{study_area}
\alias{study_area}
\title{Study area example}
\format{A \code{sf} with 4 shapefiles:}
\usage{
study_area
}
\description{
shapefile. Biosphere reserve Mariposa Monarca plus a buffer of 10 km. It is
used to run the MK_dECA and test_ECA_distance examples.
}
\references{

}
\keyword{datasets}
