% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sym_matrix.R
\name{to_sym_matrix}
\alias{to_sym_matrix}
\title{Pairwise to symmetric matrix}
\usage{
to_sym_matrix(
  df,
  from = "from",
  to = "to",
  value = "value",
  diag_value = 0,
  fill_na = 0
)
}
\arguments{
\item{df}{data.frame}

\item{from}{character}

\item{to}{character}

\item{value}{character}

\item{diag_value}{numeric}

\item{fill_na}{numeric}
}
\description{
Pairwise to symmetric matrix
}
\keyword{internal}
