% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopoClean.R
\name{TopoClean}
\alias{TopoClean}
\title{Topological error correction}
\usage{
TopoClean(x, xsimplify = FALSE)
}
\arguments{
\item{x}{Object of class sf, sfc, sfg or SpatialPolygons}

\item{xsimplify}{logical.}
}
\description{
Topological error correction
}
