% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_convert.R
\name{unit_convert}
\alias{unit_convert}
\title{Convert numeric types from one area unit to another}
\usage{
unit_convert(
  data_unit,
  unit_1 = c("m2", "Dam2", "ha", "km2", "inch2", "foot2", "yard2", "mile2", "m", "km",
    "inch", "foot", "yard", "mile"),
  unit_2 = c("m2", "Dam2", "ha", "km2", "inch2", "foot2", "yard2", "mile2", "m", "km",
    "inch", "foot", "yard", "mile")
)
}
\arguments{
\item{data_unit}{\code{numeric}. Area value to be converted.}

\item{unit_1}{\code{character}. Select an area unit: "m2" = square meter, "Dam2" = square decametre,
"ha" = hectares, "km2" = square kilometer, "inch2" = square inch, "foot2" = square foot,
"yard2" = square yard, "mile2" = square mile. Length unit: "m" = meter, "km" = kilometer, "inch", "foot", "yard", "mile"}

\item{unit_2}{\code{character}. Select an area unit: "m2" = square meter, "Dam2" = square decametre,
"ha" = hectares, "km2" = square kilometer, "inch2" = square inch, "foot2" = square foot,
"yard2" = square yard, "mile2" = square mile. Length unit: "m" = meter, "km" = kilometer, "inch", "foot", "yard", "mile"}
}
\description{
Convert numeric types from one area unit to another
}
\examples{
unit_convert(100, "km2", "ha")
}
