% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation_patches.R
\docType{data}
\name{vegetation_patches}
\alias{vegetation_patches}
\title{Cores shapefile for analysis}
\format{
A \code{sf} with 142 features and 1 field, which are:
\describe{
\item{id}{Vegetation patches identifier.}
 }
}
\usage{
vegetation_patches
}
\description{
A simple feature collection \code{sf} with 142 Vegetation patches with less alteration or degradation that
 is for testing and running of MK_fragmentation and MK_dPCIIC examples.
}
\references{
INEGI. (2013). Conjunto de datos vectoriales de uso del suelo y vegetacion, serie V (capa union), escala 1:250, 000.
Instituto Nacional de Estadistica y Geografia, Aguascalientes.
}
\keyword{datasets}
