% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{cost_distances}
\alias{cost_distances}
\title{Cost distances}
\usage{
cost_distances(x, id, type_distance = "least-cost", resistance = NULL,
  CostFun = NULL, ngh = NULL, mask = NULL, threshold = NULL,
  geometry_out = NULL, write_table = NULL)
}
\arguments{
\item{x}{object of class sf, sfc, sfg or SpatialPolygons. The shapefile must be in a projected coordinate system.}

\item{id}{character. Column name with the core id.}

\item{type_distance}{character. Choose one of the distances: "least-cost" (default) that takes into account obstacles and local friction of the landscape (See, gdistance package);
"commute-time" that is analogous to the resistance distance. This distance is based on the random walk theory and calculated using the electrical circuit theory (See, gdistance package).}

\item{resistance}{raster. Raster object with resistance values (landscape friction).}

\item{CostFun}{A function to compute the cost to move between cells.The default is the mean (isotropic cost distance):  function(x) mean(x).}

\item{ngh}{numeric.  Neighbour graph (directions) for distance calculations: 4 (von Neu-mann neighbourhood), 8 (Moore neighbourhood) or 16. Default equal **8**.}

\item{mask}{object of class sf, sfc, sfg or SpatialPolygons. For higher processing speed use this option to clip the resistance at the extent of the mask.}

\item{threshold}{numeric. Distance threshold, pairs of nodes with a distance value above this threshold will be discarded.}

\item{geometry_out}{numeric. If some spatial geometries are out of the resistance extent, then a buffer zone the large enough to cover these spatial geometries and with this numeric value will be added to the resistance, so that it is possibñe to calculate a cost distance value for the pairs of nodes that involve these geometries.
If NULL, then a Euclidean distance will be calculated to find these distances.}

\item{write_table}{character.  "" indicates output to the console.}
}
\value{
cost distance matrix
}
\description{
Cost distances
}
\references{
https://cran.r-project.org/web/packages/gdistance/gdistance.pdf
}
