% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{cost_distances}
\alias{cost_distances}
\title{Cost distances}
\usage{
cost_distances(
  x,
  id,
  type_distance = "least-cost",
  resistance = NULL,
  CostFun = NULL,
  ngh = NULL,
  mask = NULL,
  threshold = NULL,
  geometry_out = NULL,
  bounding_circles = NULL,
  pairwise = TRUE,
  write_table = NULL
)
}
\arguments{
\item{x}{object of class sf, sfc, sfg or SpatialPolygons. The shapefile must be in a projected coordinate system.}

\item{id}{character. Column name with the core id.}

\item{type_distance}{character. Choose one of the distances: "least-cost" (default) that takes into account obstacles and local friction of the landscape (See, gdistance package);
"commute-time" that is analogous to the resistance distance. This distance is based on the random walk theory and calculated using the electrical circuit theory (See, gdistance package).}

\item{resistance}{raster. Raster object with resistance values (landscape friction). Consider using a range of resistance values between 1 and 10, NA should be a complete barrier.}

\item{CostFun}{A function to compute the cost to move between cells. The default is the conductance
(isotropic cost distance):  function(x) 1/x[2].}

\item{ngh}{numeric. Neighbour graph (directions) for distance calculations: 4 (von Neu-mann neighbourhood), 8 (Moore neighbourhood) or 16 (king’s and knight’s moves). Default equal 8.}

\item{mask}{object of class sf, sfc, sfg or SpatialPolygons. For higher processing speed use this option to clip the resistance at the extent of the mask.}

\item{threshold}{numeric. Distance threshold, pairs of nodes with a distance value above this threshold will be discarded.}

\item{geometry_out}{numeric. Avoids the "Inf" error and corrects the "Inf" distance values when some spatial geometries are out
of the resistance extent. The raster NA values will be replaced by a CONDUCTANCE value that must be provided by the user
using this argument, so that it is possible to calculate cost distances for all the pairs of nodes.
 If NULL, then a Euclidean distance will be calculated to find these distances.}

\item{bounding_circles}{numeric. If a value is entered, this will create bounding circles around pairs of core areas
(recommended for speed, large resistance rasters or pixel resolution < 150 m).
Buffer distances are entered in map units. Also, the function is parallelized using
 and furrr package and multiprocess plan, default = NULL.}

\item{pairwise}{logical. If TRUE a pairwise table is returned (From, To, distance) otherwise it will be a matrix.}

\item{write_table}{character. "" indicates output to the console.}
}
\value{
cost distance matrix
}
\description{
Cost distances
}
\details{
The function builds on functions out of Jacob van Etten’s ’gdistance’ package.
}
\references{
https://cran.r-project.org/web/packages/gdistance/gdistance.pdf
}
