% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{euclidean_distances}
\alias{euclidean_distances}
\title{Euclidian distances}
\usage{
euclidean_distances(x, id, type_distance = "centroid",
  distance_unit = "m", tolerance = NULL, threshold = NULL,
  write_table = NULL)
}
\arguments{
\item{x}{object of class sf, sfc, sfg or SpatialPolygons. The shapefile must be in a projected coordinate system.}

\item{id}{character. Column name with the core id.}

\item{type_distance}{character. Choose one of the distances: "centroid" (faster, default), where Euclidean distance is calculated from feature centroid; "edge", where Euclidean distance is calculated from feature edges;}

\item{distance_unit}{character. Set a distance unit, "Makurhini::unit_covert()" compatible unit ("m", "km", "inch", "foot", "yard", "mile"). Default equal to meters "m".}

\item{tolerance}{numeric. Argument for higher processing speed. In case you have selected the "edge" distance, use this option to simplify the geometry and reduce the
number of vertices (from rgeos::gSimplify).}

\item{threshold}{numeric. Distance threshold, pairs of nodes with a distance value above this threshold will be discarded.}

\item{write_table}{character. "" indicates output to the console.}
}
\value{
Pairwise Euclidean distance table
}
\description{
Euclidian distances
}
\references{
Douglas, David H. and Peucker, Thomas K. (1973) "Algorithms for the Reduction of the Number of Points Required to Represent a Digitised Line or its Caricature", The Canadian Cartographer, 10(2), pp112-122.
}
