% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protconn_grid.R
\name{get_protconn_grid}
\alias{get_protconn_grid}
\title{Estimates ProtConn metric and fractions for a grid}
\usage{
get_protconn_grid(x, y, p, pmedian = TRUE, d, LA = NULL, bound = FALSE)
}
\arguments{
\item{x}{list.}

\item{y}{distance matrix}

\item{p}{numeric. Probability}

\item{pmedian}{logical. median (TRUE) or mean(FALSE) dispersal distance}

\item{d}{numeric. Dispersal distance}

\item{LA}{numeric. Max. landscape attribute}

\item{bound}{logical. If TRUE then ProtConn bound will be estimated}
}
\description{
Estimates ProtConn metric and fractions for a grid
}
