% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_centroid_within_poly.R
\name{make_grid}
\alias{make_grid}
\title{Make grid}
\usage{
make_grid(x, type, cell_area, clip = FALSE, tolerance = NULL,
  grid_boundary = NULL)
}
\arguments{
\item{x}{Object of class sf, sfc, sfg or SpatialPolygons}

\item{type}{character. c("square", "hexagonal")}

\item{cell_area}{numeirc. Grid area (square kilometers).}

\item{clip}{logical. If TRUE, the new grid will be clipped to the region area. The operation time will be longer the greater the number of vertices in the polygon of the region, if it is a region with many vertices use the argument "tolerance".}

\item{tolerance}{numeric. If Reduces the number of vertices in the region polygon.}

\item{grid_boundary}{logical.If TRUE, the Incomplete "hexagons" or "squares" in the boundaries of the region will be discarded}
}
\description{
Make grid
}
