% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_BCentrality.R
\name{MK_BCentrality}
\alias{MK_BCentrality}
\title{Betweenness Centrality metrics using Conefor command line}
\usage{
MK_BCentrality(
  nodes,
  id,
  attribute = NULL,
  area_unit = "ha",
  distance = list(type = "centroid", resistance = NULL),
  metric = c("BC", "BCIIC", "BCPC"),
  distance_thresholds = NULL,
  probability = NULL,
  LA = NULL,
  coneforpath = NULL,
  dA = FALSE,
  dvars = FALSE,
  parallel = NULL,
  rasterparallel = FALSE,
  write = NULL,
  intern = TRUE
)
}
\arguments{
\item{nodes}{Object containing nodes (e.g., habitat patches or fragments) information. It can be of the following classes:\cr
-   \code{Data.frame} with at least two columns: the first for node IDs and the second for attributes. \cr
-   Spatial data of type vector (class \code{sf, SpatVector, SpatialPolygonsDataFrame}). It must be in a projected coordinate system.\cr
-   Raster (class \code{RasterLayer, SpatRaster}). It must be in a projected coordinate system. The values must be integers representing the ID of each habitat patch or node, with non-habitat areas represented by NA values (see \link[raster]{clump} or \link[terra]{patches}).}

\item{id}{\code{character}. If nodes is a vector then you can specify the column name with the node ID.
If NULL, then a new id will be generated.}

\item{attribute}{\code{character} or \code{vector}. If \code{NULL} the area of the nodes will be used as the node attribute. The unit of area can be selected using the \code{area_unit} parameter. To use an alternative attribute, consider the class type of the object in the \code{nodes} parameter: \cr
-   If \code{nodes} is a spatial vector or data.frame, specify \bold{the name of the column} containing the attribute for the nodes. \cr
-   If \code{nodes} is a raster layer then it must be a numeric vector with the node's attribute. The length of the vector must be equal to the number of nodes.}

\item{area_unit}{\code{character}. (\emph{optional, default = } \code{"m2"}) \cr. A \code{character} indicating the area units when \code{attribute} is \code{NULL}. Some options are "m2" (the default), "km2", "cm2", or "ha";  See \link[Makurhini]{unit_convert} for details.}

\item{distance}{A \code{matrix} or \code{list} to establish the distance between each pair of nodes. Distance between nodes may be Euclidean distances (straight-line distance) or effective distances (cost distances) by considering the landscape resistance to the species movements. \cr
- If it is a \code{matrix}, then the number of columns and rows must be equal to the number of nodes. This distance matrix could be generated by the \link[Makurhini]{distancefile} function.\cr
- If it is a \code{list} of parameters, then it must contain the distance parameters necessary to calculate the distance between nodes. For example, two of the most important parameters: \code{“type”} and \code{“resistance”}. For \code{"type"} choose one  of the distances:  \bold{"centroid" (faster), "edge", "least-cost" or "commute-time"}. If the type is equal to \code{"least-cost"} or \code{"commute-time"}, then you must use the \code{"resistance"} argument. For example: \code{distance(type = "least-cost", resistance = raster_resistance)}. \cr
To see more arguments see the \link[Makurhini]{distancefile} function.}

\item{metric}{A \code{character} indicating the Betweenness Centrality metric to use: \code{"BC"} (the classical), \code{“BCIIC”} considering binary indices and topological distances, and \code{“BCPC”} (recommended)  considering the maximum product probabilities.}

\item{distance_thresholds}{A \code{numeric} indicating the dispersal distance or distances (meters) of the considered species. If \code{NULL} then distance is estimated as the median dispersal distance between nodes. Alternatively, the \link[Makurhini]{dispersal_distance} function can be used to estimate the dispersal distance using the species home range.}

\item{probability}{A \code{numeric} value indicating the probability that corresponds to the distance specified in the \code{distance_threshold}. For example, if the \code{distance_threshold} is a median dispersal distance, use a probability of 0.5 (50\%). If the \code{distance_threshold} is a maximum dispersal distance, set a probability of 0.05 (5\%) or 0.01 (1\%). Use in case of selecting the \code{"BCPC"} metric. If \code{probability = NULL}, then a probability of 0.5 will be used.}

\item{LA}{\code{numeric}. (\emph{optional, default = } \code{NULL}). The maximum landscape attribute, which is the attribute value that would correspond to a hypothetical habitat patch covering all the landscape with the best possible habitat, in which IIC and PC would be equal to 1. For example, if nodes attribute corresponds to the node area, then LA equals total landscape area. If nodes attribute correspond to a quality-weighted area and the quality factor ranges from 0 to 100, LA will be equal to 100 multiplied by total landscape area. The value of LA does not affect at all the importance of the nodes and is only used to calculate the overall landscape connectivity. If no LA value is entered (default) and  \code{overall = TRUE} or \code{onlyoverall = TRUE}, the function will only calculate the numerator of the global connectivity indices and the equivalent connected ECA or EC index.}

\item{coneforpath}{\code{character}. Path to Conefor 2.6 with command line interface (http://www.conefor.org/coneforsensinode.html). Example, \code{"C:/Users/coneforWin64.exe"}.}

\item{dA}{\code{logical}. If TRUE, then the delta attribute will be added to the node's importance result.}

\item{dvars}{\code{logical}. If TRUE, then the absolute variation will be added to the node's importance result.}

\item{parallel}{(\emph{optional, default =} \code{NULL}).
A \code{numeric} specifying the number of cores to parallelize the index estimation of the PC or IIC index and its deltas.Particularly useful when you have more than 1000 nodes. By default the analyses are not parallelized.}

\item{rasterparallel}{\code{logical}. If parallel is \code{FALSE} and nodes is a \code{raster} then you can use this argument to assign the metrics values to the nodes raster. It is useful when raster resolution is less than 100 m\out{<sup>2</sup>}.}

\item{write}{\code{character}. Write output shapefile, example, "C:/ejemplo.shp".}

\item{intern}{\code{logical}. Show the progress of the process, default = TRUE. Sometimes the advance process does not reach 100 percent when operations are carried out very quickly.}
}
\description{
Use this function to calculate the BC, BCIIC and BCPC indexes under one or several distance thresholds using Conefor command line.
}
\details{
To use this function  you need to download and uncompress Conefor 2.6 with command line interface from \url{http://www.conefor.org/coneforsensinode.html}\cr
Betweenness Centrality metrics can be calculated in three different ways:\cr
- \bold{BC} calculates the classical Betweenness Centrality metric as originally defined by Freeman (1977; Sociometry 40: 35–41). \cr
- \bold{BCIIC and BCPC} calculate the improved version of the BC metric by Bodin and Saura (2010) integrated within the same analytical framework as the IIC (binary) and the PC (probabilistic) metrics.
}
\examples{
\dontrun{
library(Makurhini)
data("habitat_nodes", package = "Makurhini")
nrow(habitat_nodes) #Number of nodes

#Two distance thresholds.
#You need to place your path to the conefor .exe
BCIIC <- MK_BCentrality(nodes = habitat_nodes, id = "Id",
            coneforpath = "C:/Users/coneforWin64.exe",
            distance = list(type = "centroid"),
            metric = "BCIIC", LA = NULL,
            distance_thresholds = c(10000, 30000)) #10 and 30 km

#Using raster
data("habitat_nodes_raster", package = "Makurhini")
##Using parallel
BCPC_parallel <- MK_BCentrality(nodes = habitat_nodes_raster,
                    coneforpath = "C:/Users//coneforWin64.exe",
                    id = "id", attribute = NULL,
                    distance = list(type = "centroid"),
                    metric = "BCPC", LA = NULL, probability = 0.5,
                    distance_thresholds = 40000,
                    parallel = 4) #40 and 60 km
}
}
\references{
Saura, S. and Torne, J. (2012). Conefor 2.6. Universidad Politecnica de Madrid. Available at \url{www.conefor.org}.\cr
 Freeman L.C. (1977). Set of Measures of Centrality Based on Betweenness. Sociometry 40: 35-41.\cr
 Bodin, O. and Saura, S. (2010). Ranking individual habitat patches as connectivity providers: integrating network analysis and patch removal experiments. Ecological Modelling 221: 2393-2405.
}
