% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_Connect_grid.R
\name{MK_Connect_grid}
\alias{MK_Connect_grid}
\title{Connectivity indexes in a regular grid}
\usage{
MK_Connect_grid(
  grid_pol = NULL,
  grid_id = NULL,
  grid_type = c("hexagonal", "square"),
  cellsize = NULL,
  grid_boundary = FALSE,
  clip = FALSE,
  tolerance = NULL,
  nodes,
  region = NULL,
  attribute = NULL,
  thintersect = NULL,
  area_unit = "ha",
  metric = c("ProtConn", "PC"),
  distance = list(type = "centroid"),
  distance_threshold = NULL,
  probability = NULL,
  transboundary = NULL,
  intern = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{grid_pol}{object of class sf, sfc, sfg or SpatialPolygons. Grid hexagones or squares. The
shapefile must be in a projected coordinate system.}

\item{grid_id}{character. Column name of the grid ID.}

\item{grid_type}{character. If grid is null you can make a regular grid of "hexagonal" or "square".}

\item{cellsize}{numeric. Grid area (square kilometers).}

\item{grid_boundary}{logical.If TRUE, the Incomplete "hexagons" or "squares" in the boundaries of
the region will be discarded}

\item{clip}{logical. If TRUE, the new grid will be clipped to the region area. The operation time
will be longer the greater the number of vertices in the polygon of the region,
if it is a region with many vertices use the argument "tolerance".}

\item{tolerance}{numeric. If "clip" is equal to TRUE reduces the number of vertices in the region polygon.}

\item{nodes}{object of class sf, sfc, sfg or SpatialPolygons. Nodes shapefile, the shapefile must be
in a projected coordinate system.}

\item{region}{object of class sf, sfc, sfg or SpatialPolygons. Region shapefile, the shapefile must be
in a projected coordinate system.}

\item{attribute}{character. Select the nodes attribute: "Intersected area" = Available if the metric argument
is equal to ProtConn, it corresponds to intersected Protected areas (default); or
another specific column name with the nodes attribute, ideally this attribute mus be an area-weighted index,
 otherwise the interpretation of the ProtConn or PC metric may change.}

\item{thintersect}{numeric. Only available if you selected ProtConn as metric. Threshold of intersection in percentage allowed to select or not a target geometry.
Example, if thintersect is equal to 90 then a node will be selected only if the intersection between the node and
 the region is >= 90 percentage. If NULL, thintersect will be 0 (default). (see, "Makurhini::MK_ProtConn")}

\item{area_unit}{character. Attribute area units. You can set an area unit, "Makurhini::unit_covert()" compatible unit ("m2", "Dam2, "km2", "ha", "inch2", "foot2", "yard2", "mile2"). Default equal to hectares "ha".}

\item{metric}{character. Choose a connectivity metric: "ProtConn" Protected Connected Land or "PC" Probability of conectivity considering maximum product probabilities.}

\item{distance}{list. See distancefile(). Example, list(type= "centroid", resistance = NULL).}

\item{distance_threshold}{numeric. Distance threshold to establish connections (meters).}

\item{probability}{numeric. Connection probability to the selected distance threshold, e.g., 0.5
(default) that is 50 percentage of probability connection. Use in case of selecting the "PC"
metric or "ProtConn". If probability = NULL, then it will be the inverse of the mean dispersal distance
for the species (1/α; Hanski and Ovaskainen 2000).}

\item{transboundary}{numeric. Buffer to select polygones in a second round, their attribute value = 0,
see  "Makurhini::MK_ProtConn".}

\item{intern}{logical. Show the progress of the process, default = TRUE.}

\item{parallel}{logical. Parallelize the function using furrr package and multiprocess
plan, default = FALSE.}
}
\description{
Use the function to compute the Protected Connected (ProtConn), EC, PC or IIC indexes in a regular grid.
}
\examples{
\dontrun{
library(Makurhini)
library(raster)

data("Protected_areas", package = "Makurhini")
data("regions", package = "Makurhini")
ecoregion <- regions[2,]
plot(ecoregion, col="blue")

hexagons_priority <- MK_Connect_grid(grid_type = "hexagonal",
                                    cellsize = 10000, grid_boundary = FALSE,
                                    clip = FALSE, nodes = Protected_areas, region = ecoregion,
                                    attribute = "Intersected area", thintersect = NULL,
                                    area_unit = "ha", metric = "ProtConn",
                                    distance = list(type = "centroid"),
                                    distance_threshold = 3000,
                                    probability = 0.5, transboundary = 6000,
                                    intern = TRUE, parallel = FALSE)
hexagons_priority
plot(hexagons_priority["ProtConn"])
}
}
\references{
Matt Strimas-Mackey. \url{http://strimas.com/spatial/hexagonal-grids/}.\cr
Saura, S., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2017). Protected areas in the
world's ecoregions: How well connected are they? Ecological Indicators, 76, 144–158.
Saura, S. & Torne, J. (2012). Conefor 2.6. Universidad Politecnica de Madrid. Available
 at \url{www.conefor.org}.\cr
Pascual-Hortal, L. & Saura, S. (2006). Comparison and development of new graph-based landscape
 connectivity indices: towards the priorization of habitat patches and corridors for conservation.
 Landscape Ecology, 21(7): 959-967.\cr
Saura, S. & Pascual-Hortal, L. (2007). A new habitat availability index to integrate connectivity
in landscape conservation planning: comparison with existing indices and application to a case study.
Landscape and Urban Planning, 83(2-3): 91-103.
}
