% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_dECA.R
\name{MK_dECA}
\alias{MK_dECA}
\title{ECA, dA and dECA.}
\usage{
MK_dECA(
  nodes,
  attribute = NULL,
  area_unit = "m2",
  distance = list(type = "centroid", resistance = NULL),
  metric = "IIC",
  probability = NULL,
  distance_thresholds = NULL,
  LA = NULL,
  plot = FALSE,
  parallel = NULL,
  write = NULL,
  intern = TRUE
)
}
\arguments{
\item{nodes}{list of objects class sf, SpatialPolygonsDataFrame or raster. Nodes of each time to analyze.
The shapefiles must be in a projected coordinate system.}

\item{attribute}{character or vector. If nodes is a shappefile then you must specify the column name
with the attribute in the data table selected for the nodes. If nodes is a raster layer then it must be
a numeric vector with the node's attribute. The length of the vector must be equal to the number of nodes.
 The numeric vector is multiplied by the area of each node to obtain a weighted habitat index.
 If NULL the node area will be used as a node attribute, the unit area can be selected using the "area_unit" argument.}

\item{area_unit}{character. If attribute is NULL you can set an area unit, "Makurhini::unit_covert()"
compatible unit(e.g., "m2", "km2", "ha"). Default equal to hectares "m2".}

\item{distance}{list. Distance parameters. For example: type, resistance,or keep. For "type" choose one of the
distances: "centroid" (faster), "edge", "least-cost" or "commute-time". If the type is equal to "least-cost"
or "commute-time", then you have to use the "resistance" argument. You can place a list with resistances
where there must be one resistance for each time/scenario of patches in the **nodes** parameter, for example,
if nodes has a list with two time patches then you can use two resistances one for time 1 and one for time 2:
distance(type = "least-cost", resistance = list(resistanceT1, resistanceT2)). To See more arguments consult
the help function of distancefile().}

\item{metric}{character. Choose a connectivity metric: "IIC" considering topologycal distances or "PC"
considering maximum product probabilities.}

\item{probability}{numeric. Connection probability to the selected distance threshold, e.g.,
0.5 that is 50 percentage of probability connection. Use in case of selecting the "PC" metric.
If probability = NULL, then it will be the inverse of the mean dispersal distance for the species (1/α; Hanski and Ovaskainen 2000).}

\item{distance_thresholds}{numeric. Distance or distances thresholds to establish connections (meters).
For example, distance_threshold = 30000; two or more specific distances:
distance_threshold = c(30000, 50000); sequence distances: distance_threshold = seq(10000,100000, 10000).}

\item{LA}{numeric. Maximum landscape attribute ("units" equal to "area_unit", default equal to "ha").}

\item{plot}{logical. Also, you can provide the corresponding year for each period of
time analyzed, e.g., c("2011", "2014", "2017")}

\item{parallel}{numeric. Specify the number of cores to use for parallel processing, default = NULL.
Parallelize the function using furrr package and multiprocess plan.}

\item{write}{character. Path and name of the output ".csv" file}

\item{intern}{logical. Show the progress of the process, default = TRUE. Sometimes the advance process does not reach 100 percent when operations are carried out very quickly.}
}
\value{
Table with:\cr
A: Area in km2\cr
ECA: ECA value\cr
Normalized_ECA: Relative connectivity (percentage)\cr
dA: Delta Area between times (percentage)\cr
dECA: Delta ECA between times (percentage)\cr
Type_change: Type of change using the dECAfun() and the difference between dA and dECA.\cr
}
\description{
Equivalent Connected Area (ECA; if the area is used as attribute) or Equivalent Connectivity index (EC)
}
\examples{
\dontrun{
library(Makurhini)
library(sf)

data("list_forest_patches", package = "Makurhini")
data("study_area", package = "Makurhini")
class(list_forest_patches)

Max_attribute <- unit_convert(st_area(study_area), "m2", "ha")

dECA_test <- MK_dECA(nodes= list_forest_patches, attribute = NULL, area_unit = "ha",
                  distance = list(type= "centroid"), metric = "PC",
                  probability = 0.05, distance_thresholds = 5000,
                  LA = Max_attribute, plot= c("1993", "2003", "2007", "2011"),
                  intern = TRUE)
dECA_test

}
}
\references{
\url{www.conefor.org}\cr
Saura, S., Estreguil, C., Mouton, C., & Rodríguez-Freire, M. (2011). Network analysis to assess landscape connectivity trends: Application to European forests (1990-2000). Ecological Indicators, 11(2), 407–416.
https://doi.org/10.1016/j.ecolind.2010.06.011 \cr
Herrera, L. P., Sabatino, M. C., Jaimes, F. R., & Saura, S. (2017). Landscape connectivity and the role of small habitat patches as stepping stones: an assessment of the grassland biome in South America. Biodiversity and Conservation, 26(14), 3465–3479.
https://doi.org/10.1007/s10531-017-1416-7
}
