% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_dPCIIC.R
\name{MK_dPCIIC}
\alias{MK_dPCIIC}
\title{Estimate the integral index of connectivity (IIC) or the probability of connectivity (PC)}
\usage{
MK_dPCIIC(nodes, id = NULL, attribute = NULL, area_unit = "ha",
  restauration = NULL, distance = list(type = "centroid", resistance =
  NULL), metric = c("IIC", "PC"), probability,
  distance_thresholds = NULL, overall = FALSE, dA = FALSE,
  dvars = FALSE, LA = NULL, write = NULL)
}
\arguments{
\item{nodes}{Object of class sf, sfc, sfg or SpatialPolygons.The shapefile must be in a projected coordinate system.}

\item{id}{character. Column name with the nodes id.If NULL, then a new temporal id will be generated.}

\item{attribute}{character. Column name with the nodes attribute. If NULL, then the patch area (ha) will be estimated and used as the attribute.}

\item{area_unit}{character. If attribute is NULL you can set an area unit, udunits2 package compatible unit (e.g., "km2", "cm2", "ha"). Default equal to square meters "ha".}

\item{restauration}{character. Name of the column with restauration value. Binary values (0,1), where 1 = existing nodes in the landscape, and 0 = a new node to add to the initial landscape (restored).}

\item{distance}{list. Distance parameters. For example: type, resistance,or tolerance. For "type" choose one of the distances: "centroid" (faster), "edge", "hausdorff-edge",
"least-cost" or "commute-time". If the type is equal to "least-cost" or "commute-time", then you have to use the "resistance" argument.
  To See more arguments consult the help function of distancefile().}

\item{metric}{character. Choose a connectivity metric: "IIC" considering topologycal distances or "PC" considering maximum product probabilities.}

\item{probability}{numeric. Connection probability to the selected distance threshold, e.g., 0.5 that is 50 percentage of probability connection. Use in case of selecting the "PC" metric.}

\item{distance_thresholds}{numeric. Distance or distances thresholds to establish connections. For example, one distance: distance_threshold = 30000; two or more specific distances:
distance_thresholds = c(30000, 50000); sequence distances: distance_thresholds = seq(10000,100000, 10000).}

\item{overall}{logical. If TRUE, then the EC index will be added to the result which is transformed into a list. Default equal to FALSE}

\item{dA}{logical. If TRUE, then the delta attribute will be added to the node's importance result.}

\item{dvars}{logical. If TRUE, then the absolute variation will be added to the node's importance result.}

\item{LA}{numeric. Maximum landscape attribute (attribute unit, if attribute is NULL then unit is equal to ha).}

\item{write}{character. Write output shapefile and overall table (if TRUE overall argument).
It is necessary to specify the "Folder direction" + "Initial prefix",  for example, "C:/ejemplo".}
}
\description{
Use this function to calculate the PC and IIC indexes under one or several distance thresholds.
}
\examples{
ruta <- system.file("extdata", "Fragmentation.RData", package = "Makurhini")
load(ruta)
nrow(cores)
#One distance threshold IIC
IIC <- MK_dPCIIC(nodes = cores, id = "id", attribute = NULL,
                    distance = list(type = "centroid"),
                    metric = "IIC", distance_thresholds = 30000)
IIC
#Two or more distance thresholds PC
PC <- MK_dPCIIC(nodes = cores, id = "id", attribute = NULL,
                    distance = list(type = "centroid"),
                    metric = "PC", probability = 0.5,
                    distance_thresholds = c(5000, 10000))
PC
}
\references{
Saura, S. & Torné, J. 2012. Conefor 2.6 user manual (May 2012). Universidad Politécnica de Madrid. Available at \url{www.conefor.org}.\cr
Pascual-Hortal, L. & Saura, S. 2006. Comparison and development of new graph-based landscape connectivity indices: towards the priorization of habitat patches and corridors for conservation. Landscape Ecology 21 (7): 959-967.\cr
Saura, S. & Pascual-Hortal, L. 2007. A new habitat availability index to integrate connectivity in landscape conservation planning: comparison with existing indices and application to a case study. Landscape and Urban Planning 83 (2-3): 91-103.
}
