% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_Fragmentation.R
\name{MK_Fragmentation}
\alias{MK_Fragmentation}
\title{Fragmentation Statistics}
\usage{
MK_Fragmentation(patches, edge_distance = 500, min_patch_area = 100,
  landscape_area = NULL, area_unit = "km2", perimeter_unit = "km",
  plot = FALSE, write = NULL)
}
\arguments{
\item{patches}{Object of class sf, sfc, sfg or SpatialPolygons with individual patches. The shapefile must be in a projected coordinate system.}

\item{edge_distance}{Numeric. Distance to edge in meters. Default equal 500 m (Haddad et al. 2015)}

\item{min_patch_area}{Numeric. Minimum patch area. Default equal 100 km2(Haddad et al. 2015)}

\item{landscape_area}{Numeric. Total landscape area in km2 (optional). If NULL the total patch area will be used.}

\item{area_unit}{character. You can set an area unit (e.g., "km2", "cm2", "m2", "ha"; see Makurhini::unit_convert). Default equal to square kilometers "km2".}

\item{perimeter_unit}{character. You can set a perimeter unit (e.g., "km", "cm", "m", "ha"; see Makurhini::unit_convert). Default equal to kilometers "km".}

\item{plot}{Logical. Basic histograms and core area - edge map.}

\item{write}{Character. Write the following outputs: Fragmentation.csv, Fragmentation.shp and plots. It's necessary to specify the path and prefix, for example: "C:/Folder/Fragmentation".}
}
\value{
Patch and landscape statistics:\cr
1) Patches Area in square kilometers.\cr
2) Number patches.\cr
3) Mean size of patches.\cr
4) Number of patches smaller than 100 km2 or another minimum patch area.\cr
5) Area in square kilometers of patches smaller than 100 km2 or another minimum patch area.\cr
6) Total edge.\cr
7) Edge density.\cr
8) Total core area (km2; considering a distance to edge of 500 m).\cr
9) Total core area.\cr
10) Core percent.\cr
11) Edge percent.\cr
12) Cority index. It is a measure of fragmentation with respect to a distance from the core area of 500 m (Haddad et al., 2015), where a value of 1 indicates a landscape without fragmentation.\cr
13) Shape Index. A simple shape metric that takes values from 1 (perfectly compact) to infinity is derived by dividing the perimeter by the perimeter of a circle of the same area.\cr
14) Fractal dimension.\cr
15) Effective Mesh Size.\cr
}
\description{
Calculate patch and landscape statistics
}
\examples{
data("vegetation_patches", package = "Makurhini")
nrow(vegetation_patches) # Number of patches
fragmentation <- MK_Fragmentation(patches = vegetation_patches, edge_distance = 500, plot = TRUE)

#Table
fragmentation$`Summary landscape metrics (Viewer Panel)`
#Shapefile
fragmentation$`Patch statistics shapefile`
}
\references{
McGarigal, K., S. A. Cushman, M. C. Neel, and E. Ene. 2002. FRAGSTATS: Spatial Pattern Analysis Program for Categorical Maps. Computer software program produced by the authors at the University of Massachusetts, Amherst. Available at the following web site:
 \url{www.umass.edu/landeco/research/fragstats/fragstats.html}.\cr
Haddad et al. (2015). Science Advances 1(2):e1500052. \url{DOI: 10.1126/sciadv.1500052}.
}
