% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_ProtConn.R
\name{MK_ProtConn}
\alias{MK_ProtConn}
\title{Protected Connected (ProtConn)}
\usage{
MK_ProtConn(
  nodes = NULL,
  region = NULL,
  area_unit = "m2",
  distance = list(type = "centroid", resistance = NULL),
  distance_thresholds = NULL,
  probability = 0.5,
  transboundary = NULL,
  transboundary_type = "nodes",
  protconn_bound = FALSE,
  geom_simplify = FALSE,
  delta = FALSE,
  plot = FALSE,
  write = NULL,
  parallel = NULL,
  intern = TRUE
)
}
\arguments{
\item{nodes}{object of class \code{sf, sfc, sfg, spatialPolygonsDataFrame}. Spatial data of vector type that normally contains the spatial limits of protected areas. It must be in a projected coordinate system.}

\item{region}{object of class \code{sf, sfc, sfg, spatialPolygonsDataFrame}. Polygon delimiting the region or study area. It must be in a projected coordinate system.}

\item{area_unit}{\code{character}. (\emph{optional, default = } \code{"m2"}) \cr. A \code{character} indicating the area units when \code{attribute} is \code{NULL}. Some options are "m2" (the default), "km2", "cm2", or "ha";  See \link[Makurhini]{unit_convert} for details.}

\item{distance}{A \code{list} of parameters to establish the distance between each pair of nodes. Distance between nodes may be Euclidean distances (straight-line distance) or effective distances (cost distances) by considering the landscape resistance to the species movements. \cr
 This list must contain the distance parameters necessary to calculate the distance between nodes. For example, two of the most important parameters: \code{“type”} and \code{“resistance”}. For \code{"type"} choose one  of the distances:  \bold{"centroid" (faster), "edge", "least-cost" or "commute-time"}. If the type is equal to \code{"least-cost"} or \code{"commute-time"}, then you must use the \code{"resistance"} argument. For example: \code{distance(type = "least-cost", resistance = raster_resistance)}. \cr
To see more arguments see the \link[Makurhini]{distancefile} function.}

\item{distance_thresholds}{A \code{numeric} indicating the dispersal distance or distances (meters) of the considered species. If \code{NULL} then distance is estimated as the median dispersal distance between nodes. Alternatively, the \link[Makurhini]{dispersal_distance} function can be used to estimate the dispersal distance using the species home range. Can be the same length as the \code{distance_thresholds} parameter.}

\item{probability}{A \code{numeric} value indicating the probability that corresponds to the distance specified in the \code{distance_threshold}. For example, if the \code{distance_threshold} is a median dispersal distance, use a probability of 0.5 (50\%). If the \code{distance_threshold} is a maximum dispersal distance, set a probability of 0.05 (5\%) or 0.01 (1\%). Use in case of selecting the \code{"PC"} metric. If \code{probability = NULL}, then a probability of 0.5 will be used.}

\item{transboundary}{\code{numeric}. Buffer to select polygons (e.g., PAs) in a second round. The selected polygons will have an attribute value = 0, i.e., their contribution for connectivity would be as stepping stones (Saura et al. 2017). One cross-border value or one for each threshold distance can be set.}

\item{transboundary_type}{\code{character}. Two options: \code{"nodes" (methodology from Saura et al. 2017)} or \code{"region"}.\cr
- If it is \code{"nodes"}, the transboundary is built from the limits of the nodes present in the region (default).
- If it is \code{"region"}, is selected the transboundary is built from the limits of the region.}

\item{protconn_bound}{\code{logical}. If \code{TRUE} then the fractions ProtUnConn[design] and ProtConn[bound] will be estimated.}

\item{geom_simplify}{\code{logical}. Slightly simplify the region and nodes geometries.}

\item{delta}{\code{logical}. Estimate the contribution of each node to the ProtConn value in the region.}

\item{plot}{\code{logical}. Plot the main ProtConn indicators and fractions, default = \code{FALSE}.}

\item{write}{\code{character}. Output folder including the output file name without extension, e.g., \code{"C:/ProtConn/Protfiles"}.}

\item{parallel}{\code{numeric}. Specify the number of cores to use for parallel processing, default = NULL. Parallelize the function using furrr package and multiprocess
plan when there are more than one transboundary.}

\item{intern}{\code{logical}. Show the progress of the process, default = TRUE. Sometimes the advance process does not reach 100 percent when operations are carried out very quickly.}
}
\value{
Table with the following ProtConn values: \code{ECA, Prot, ProtConn, ProtUnconn, RelConn, ProtUnConn[design], ProtConn[bound], ProtConn[Prot], ProtConn[Within],
 ProtConn[Contig], ProtConn[Trans], ProtConn[Unprot], ProtConn[Within][land], ProtConn[Contig][land],
 ProtConn[Unprot][land], ProtConn[Trans][land]} \cr
\cr
- If plot \bold{is not NULL} a \code{list} is returned with the ProtConn table and a plots.
- If \bold{delta} is \code{TRUE} then it returns an sf class object with the importance value (contribution to ProtConn) for each node in the region.
}
\description{
This function calculates the Protected Connected indicator (ProtConn) for a region, its fractions and the importance (contribution) of each protected area to maintain connectivity in the region under one or more distance thresholds.
}
\examples{
\dontrun{
library(Makurhini)
library(sf)

data("Protected_areas", package = "Makurhini")
data("Ecoregions", package = "Makurhini")
region <- Ecoregions[1,]

test <- MK_ProtConn(nodes = Protected_areas, region = region,
                    area_unit = "ha",
                    distance = list(type= "centroid"),
                    distance_thresholds = c(50000, 10000),
                    probability = 0.5, transboundary = 50000,
                    plot = TRUE, parallel = NULL,
                    protconn_bound = TRUE,
                    delta = TRUE,
                    write = NULL, intern = TRUE)
test
}
}
\references{
Saura, S., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2017). Protected areas in the world’s ecoregions: How well connected are they? Ecological Indicators, 76, 144–158.\cr
Saura, S., Bertzky, B., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2018). Protected area connectivity: Shortfalls in global targets and country-level priorities. Biological Conservation, 219(October 2017), 53–67.
}
