% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_ProtConn_raster.R
\name{MK_ProtConn_raster}
\alias{MK_ProtConn_raster}
\title{Protected Connected (ProtConn) using raster nodes}
\usage{
MK_ProtConn_raster(
  nodes,
  region,
  area_unit = "m2",
  distance = list(type = "edge", resistance = NULL),
  distance_thresholds,
  probability,
  transboundary = NULL,
  transboundary_type = "nodes",
  protconn_bound = FALSE,
  delta = FALSE,
  plot = FALSE,
  resample_raster = NULL,
  write = NULL,
  parallel = NULL,
  intern = TRUE
)
}
\arguments{
\item{nodes}{Object of class  \code{raster, rast}. The file must have a projected coordinate system.}

\item{region}{object of class \code{raster, rast, sf, sfc, sfg,  SpatVector, spatialPolygonsDataFrame}. Polygon delimiting the region or study area. It must be
in a projected coordinate system.}

\item{area_unit}{\code{character}. (\emph{optional, default = } \code{"m2"}) \cr. A \code{character} indicating the area units when \code{attribute} is \code{NULL}. Some options are "m2" (the default), "km2", "cm2", or "ha";  See \link[Makurhini]{unit_convert} for details.}

\item{distance}{A list of parameters to calculate distances between node pairs.
 Supported types are \code{"edge"}, \code{"least-cost"}, and \code{"commute-time"}.
 For \code{"least-cost"} and \code{"commute-time"}, the \code{resistance} argument is required.
 Examples:
 \itemize{
   \item \code{list(type = "edge", keep = 0.5)}  # uses 50% of vertices to simplify shapes and speed up
   \item \code{list(type = "least-cost", resistance = raster_resistance)}
 }
See more arguments in \link[Makurhini]{distancefile}.}

\item{distance_thresholds}{A \code{numeric} indicating the dispersal distance or distances (meters) of the considered species. If \code{NULL} then distance is estimated as the median dispersal distance between nodes. Alternatively, the \link[Makurhini]{dispersal_distance} function can be used to estimate the dispersal distance using the species home range. Can be the same length as the \code{distance_thresholds} parameter.}

\item{probability}{A \code{numeric} value indicating the probability that corresponds to the distance specified in the \code{distance_threshold}. For example, if the \code{distance_threshold} is a median dispersal distance, use a probability of 0.5 (50\%). If the \code{distance_threshold} is a maximum dispersal distance, set a probability of 0.05 (5\%) or 0.01 (1\%). Use in case of selecting the \code{"PC"} metric. If \code{probability = NULL}, then a probability of 0.5 will be used.}

\item{transboundary}{\code{numeric}. Buffer to select polygons (e.g., PAs) in a second round. The selected polygons will have an attribute value = 0, i.e., their contribution for connectivity would be as stepping stones (Saura et al. 2017). One cross-border value or one for each threshold distance can be set.}

\item{transboundary_type}{\code{character}. Two options: \code{"nodes" (methodology from Saura et al. 2017)} or \code{"region"}.\cr
- If it is \code{"nodes"}, the transboundary is built from the limits of the nodes present in the region (default).
- If it is \code{"region"}, is selected the transboundary is built from the limits of the region.}

\item{protconn_bound}{\code{logical}. If \code{TRUE} then the fractions ProtUnConn[design] and ProtConn[bound] will be estimated.}

\item{delta}{\code{logical}. Estimate the contribution of each node to the ProtConn value in the region.}

\item{plot}{\code{logical}. Plot the main ProtConn indicators and fractions, default = FALSE.}

\item{resample_raster}{\code{numeric}. Use this parameter if you work with small high-resolution raster (e.g., less than 150 m) or with very large extents. The value of this parameter is used to resample the raster region, perform buffering and speed up the process of transboundary node selection.}

\item{write}{\code{character}. Output folder including the output file name without extension, e.g., \code{"C:/ProtConn/Protfiles"}.}

\item{parallel}{\code{numeric}. Specify the number of cores to use for parallel processing, default = NULL. Parallelize the function using furrr package and multiprocess
plan when there are more than one transboundary.}

\item{intern}{\code{logical}. Show the progress of the process, \code{default = TRUE}. Sometimes the advance process does not reach 100 percent when operations are carried out very quickly.}
}
\value{
Table with the following ProtConn values: \code{ECA, Prot, ProtConn, ProtUnconn, RelConn, ProtUnConn[design], ProtConn[bound], ProtConn[Prot], ProtConn[Within],
 ProtConn[Contig], ProtConn[Trans], ProtConn[Unprot], ProtConn[Within][land], ProtConn[Contig][land],
 ProtConn[Unprot][land], ProtConn[Trans][land]} \cr
\cr
- If plot \bold{is not NULL} a \code{list} is returned with the ProtConn table and a plots.
- If \bold{delta} is \code{TRUE} then it returns an sf class object with the importance value (contribution to ProtConn) for each node in the region.
}
\description{
Estimate Protected Connected (ProtConn) indicator and fractions for one region.
}
\examples{
\dontrun{
library(Makurhini)
library(raster)
data("Protected_areas_raster", package = "Makurhini")
plot(Protected_areas_raster)
data("Ecoregions", package = "Makurhini")
region <- Ecoregions[2,]
test <- MK_ProtConn_raster(nodes = Protected_areas_raster,
                           region = region,
                           area_unit = "ha",
                           distance = list(type= "edge", keep = 0.5),
                           distance_thresholds = c(50000, 10000),
                           probability = 0.5, transboundary = 50000,
                           plot = FALSE, parallel = NULL,
                           write = NULL, intern = FALSE)
test

}
}
\references{
Saura, S., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2017). Protected areas in the world’s ecoregions: How well connected are they? Ecological Indicators, 76, 144–158.\cr
Saura, S., Bertzky, B., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2018). Protected area connectivity: Shortfalls in global targets and country-level priorities. Biological Conservation, 219(October 2017), 53–67.
}
