% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_ProtConnMult.R
\name{MK_ProtConnMult}
\alias{MK_ProtConnMult}
\title{Multiple Protected Connected (ProtConn)}
\usage{
MK_ProtConnMult(nodes, regions, thintersect = NULL,
  attribute = "Intersected area", area_unit = "ha",
  res_attribute = 150, distance = list(type = "centroid", resistance =
  NULL), distance_thresholds, probability, transboundary = NULL,
  keep = 0.02, CI = "all", plot = FALSE, write = NULL,
  intern = TRUE, parallel = FALSE)
}
\arguments{
\item{nodes}{object of class sf, sfc, sfg or SpatialPolygons. Protected areas (PAs) shapefile.}

\item{regions}{object of class sf, sfc, sfg or SpatialPolygons. Ecoregions shapefile.}

\item{thintersect}{numeric.Threshold of intersection in percentage allowed to select or not a target geometry. Default = 90, if intersection >=90 percentage, the geometry will be selected.}

\item{attribute}{character. Select the nodes attribute: "Area" = Complete Protected areas; "Intersected area" = Intersected Protected areas; or another specific column name with the nodes attribute, ideally this attribute mus be an area-weighted index, otherwise the interpretation of the protconn index may change.}

\item{area_unit}{character. Attribute area units. You can set an area unit, udunits2 package compatible unit
(e.g., "km2", "cm2", "ha"). Default equal to hectares "ha".}

\item{res_attribute}{numeric. If the attribute is no equal to "Area" or "Intersected area" then nodes will be converted to raster to extract values in one  process step, you can set the raster resolution, default = 150.}

\item{distance}{list.See distancefile(). E.g.: list(type = "centroid", resistance = NULL, geometry_out = NULL).}

\item{distance_thresholds}{numeric. Distance or distances thresholds to establish connections. For example, one distance: distance_threshold = 30000; two or more specific distances:
distance_threshold = c(30000, 50000); sequence distances: distance_threshold = seq(10000,100000, 10000).}

\item{probability}{numeric. Connection probability to the selected distance threshold, e.g., 0.5 that is 50 percentage of probability connection. Use in case of selecting the "PC" metric.}

\item{transboundary}{numeric. Buffer to select polygones in a second round, their attribute value = 0, see Saura et al. 2017.}

\item{keep}{numeric. Simplification of the region's geometry to speed the select by location ("Makurhini::MK_selectbyloc").
The value can range from 0 to 1 and is the proportion of points to retain (default 0.02). The higher the value,
 the higher the speed but the greater uncertainty.}

\item{CI}{character. A character vector representing the type of confidence intervals that will be estimated. The value should be any subset of the values c("norm","basic", "stud", "perc", "bca") or "all" which will compute all five types of intervals (see, boot::boot.ci())}

\item{plot}{logical. Plot the main ProtConn indicators and fractions with their standard deviation, default = FALSE.}

\item{write}{character. Output folder including the output file name without extension, e.g., "C:/ProtConn/Protfiles".}

\item{intern}{logical. Show the progress of the process, default = TRUE.}

\item{parallel}{logical. Parallelize the function using furrr package and multiprocess plan, default = FALSE.}
}
\value{
Table with the following ProtConn values: ECA, Prot, ProtConn, ProtUnconn, RelConn, ProtConn[design], ProtConn[bound], ProtConn[Prot], ProtConn[Within],
ProtConn[Contig], ProtConn[Trans], ProtConn[Unprot], ProtConn[Within][land], ProtConn[Contig][land],
ProtConn[Unprot][land], ProtConn[Trans][land] \cr
\cr
*Each indicator is accompanied by six dispersion statistics: standard deviation, standard error and four confidence intervals obtained with a bootstrap-type resampling (Carpenter & Bithell 2000; see, \strong{ProtConnStat()})\cr
\cr
*If plot is not NULL a list is returned with the ProtConn table and a plots.
}
\description{
Use the CONEFOR command line to estimate Protected Connected (ProtConn) indexes for multiple ecoregions
}
\references{
Saura, S., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2017). Protected areas in the world’s ecoregions: How well connected are they? Ecological Indicators, 76, 144–158.
Saura, S., Bertzky, B., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2018). Protected area connectivity: Shortfalls in global targets and country-level priorities. Biological Conservation, 219(October 2017), 53–67.
}
