% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_ProtConnMult.R
\name{MK_ProtConnMult}
\alias{MK_ProtConnMult}
\title{Multiple Protected Connected (ProtConn)}
\usage{
MK_ProtConnMult(
  nodes,
  regions,
  thintersect = NULL,
  attribute = "Intersected area",
  area_unit = "ha",
  res_attribute = 150,
  distance = list(type = "centroid", resistance = NULL),
  distance_thresholds,
  probability,
  transboundary = NULL,
  keep = 1,
  CI = "all",
  plot = FALSE,
  write = NULL,
  intern = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{nodes}{object of class sf, sfc, sfg or SpatialPolygons. Protected areas (PAs) shapefile.}

\item{regions}{object of class sf, sfc, sfg or SpatialPolygons. Ecoregions shapefile.}

\item{thintersect}{numeric.Threshold of intersection in percentage allowed to select or not a target geometry. Default = 90, if intersection >=90 percentage, the geometry will be selected.}

\item{area_unit}{character. Attribute area units. You can set an area unit, "Makurhini::unit_covert()" compatible unit ("m2", "Dam2, "km2", "ha", "inch2", "foot2", "yard2", "mile2"). Default equal to hectares "ha".}

\item{distance}{list. See \link[Makurhini]{distancefile}. Example, list(type= "centroid", resistance = NULL).}

\item{distance_thresholds}{numeric. Distance or distances thresholds to establish connections (meters). For example, one distance: distance_threshold = 30000; two or more specific distances:
distance_threshold = c(30000, 50000); sequence distances: distance_threshold = seq(10000,100000, 10000).}

\item{probability}{numeric. Connection probability to the selected distance threshold, e.g., 0.5 that is 50 percentage of probability connection. numeric.
If probability = NULL, then it will be the inverse of the mean dispersal distance for the species (1/α; Hanski and Ovaskainen 2000).}

\item{transboundary}{numeric. Buffer to select polygones in a second round, their attribute value = 0, see Saura et al. 2017.}

\item{CI}{character. A character vector representing the type of confidence intervals that will be estimated. The value should be any subset of the values c("norm","basic", "stud", "perc", "bca") or "all" which will compute all five types of intervals (see, \link[boot]{boot.ci})}

\item{plot}{logical. Plot the main ProtConn indicators and fractions with their standard deviation, default = FALSE.}

\item{write}{character. Output folder including the output file name without extension, e.g., "C:/ProtConn/Protfiles".}

\item{intern}{logical. Show the progress of the process, default = TRUE.}

\item{parallel}{logical. Parallelize the function using furrr package and multiprocess plan, default = FALSE.}
}
\value{
Table with the following ProtConn values: ECA, Prot, ProtConn, ProtUnconn, RelConn, ProtConn[design], ProtConn[bound], ProtConn[Prot], ProtConn[Within],
ProtConn[Contig], ProtConn[Trans], ProtConn[Unprot], ProtConn[Within][land], ProtConn[Contig][land],
ProtConn[Unprot][land], ProtConn[Trans][land] \cr
\cr
*Each indicator is accompanied by six dispersion statistics: standard deviation, standard error and four confidence intervals obtained with a bootstrap-type resampling (Carpenter & Bithell 2000; see, \strong{ProtConnStat()})\cr
\cr
*If plot is not NULL a list is returned with the ProtConn table and a plots.
}
\description{
Estimate Protected Connected (ProtConn) indicator and fractions for multiple regions.
}
\examples{
\dontrun{
library(Makurhini)
library(raster)
data("Protected_areas", package = "Makurhini")
plot(Protected_areas, col="green")

data("regions", package = "Makurhini")
plot(regions, col=c("blue", "red", "green"))

test <- MK_ProtConnMult(nodes = Protected_areas, regions = regions,
                        area_unit = "ha",
                        distance = list(type= "centroid"),
                        distance_thresholds = 10000,
                        probability = 0.5, transboundary = 50000,
                        plot = TRUE)
test
}
}
\references{
Saura, S., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2017). Protected areas in the world’s ecoregions: How well connected are they? Ecological Indicators, 76, 144–158.
Saura, S., Bertzky, B., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2018). Protected area connectivity: Shortfalls in global targets and country-level priorities. Biological Conservation, 219(October 2017), 53–67.
}
