% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ECA_distance.R
\name{probability_distance}
\alias{probability_distance}
\title{Plot probability  of dispersal}
\usage{
probability_distance(probability, max_distance, eval_distance = NULL,
  min.prob = 0.2)
}
\arguments{
\item{probability}{numeric. Probability of dispersal at max_distance. If NULL}

\item{max_distance}{numeric. Up to five maximum dispersal distance (km)}

\item{eval_distance}{numeric. Calculate the probability of dispersal at a specific distance (km)}

\item{min.prob}{numeric. Value between 0-1, maximum x axe value}
}
\description{
Negative exponential dispersal kernel to calculate the probability of dispersal between two nodes. Used in probabilistic indexes (e.g., PC, BCPC, ProtConn etc.)
}
\examples{
\dontrun{
probability_distance(probability= 0.5, max_distance = c(1, 10, 30, 100))
probability_distance(probability= 0.05, max_distance = 100, eval_distance = 30)
}
}
