% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_distance.R
\name{probability_distance}
\alias{probability_distance}
\title{Plot probability  of dispersal}
\usage{
probability_distance(
  probability,
  median_distance,
  eval_distance = NULL,
  min.prob = 0.01
)
}
\arguments{
\item{probability}{numeric. Probability of dispersal at median_distance. If NULL}

\item{median_distance}{numeric. Up to five maximum dispersal distance (km)}

\item{eval_distance}{numeric. Calculate the probability of dispersal at a specific median distances (km). Available when only one median_distance is used.}

\item{min.prob}{numeric. Value between 0-1, maximum x axe value}
}
\description{
Negative exponential dispersal kernel to calculate the probability of dispersal between two nodes. Used in probabilistic indexes (e.g., PC, BCPC, ProtConn etc.)
}
\examples{
\dontrun{
probability_distance(probability= 0.5, median_distance = c(1, 10, 30, 100), min.prob = 0.01)
probability_distance(probability= 0.5, median_distance = 100, eval_distance = 100)
probability_distance(probability= 0, median_distance = 100, eval_distance = 1)
}
}
