% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Protected_areas.R
\docType{data}
\name{Protected_areas}
\alias{Protected_areas}
\title{Protected areas shapefile for analysis}
\format{A \code{SpatialPolygonsDataFrame} with 1192 features and 5 fields, which are:
\describe{
\item{OBJECTID}{Protected areas identifier.}
 }}
\usage{
Protected_areas
}
\description{
A Spatial Polygons DataFrame \code{SpatialPolygonsDataFrame} with 1192 polygons of Protected Areas of Mexico.
}
\references{
UNEP-WCMC (2019). Protected Area Profile for Mexico from the World Database of Protected Areas, May 2019.
Available at: www.protectedplanet.net
}
\keyword{datasets}
