% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Protected_areas_raster}
\alias{Protected_areas_raster}
\title{Protected areas raster for analysis}
\format{
An object of class \code{RasterLayer} of dimension 2536 x 2027 x 1.
}
\usage{
Protected_areas_raster
}
\description{
raster \code{raster} with with 957 protected areas (PAs) in these ecoregions of Colombia.
}
\references{
UNEP-WCMC. Manual de Usuario para la Base de Datos Mundial Sobre Áreas Protegidas y Base de Datos Mundial Sobre Otras Medidas Eficaces de Conservación Basadas en Áreas: 1.6; UNEP-WCMC: Cambridge, UK, 2020
}
\keyword{datasets}
